--- docs/assets/Playground-4eb268db.js
@@ -1,11 +1,0 @@
-import{r as a,W as L,j as H}from"./index-10388f2a.js";function be(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}function oe(e,t){var r=Object.keys(e);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(e);t&&(n=n.filter(function(o){return Object.getOwnPropertyDescriptor(e,o).enumerable})),r.push.apply(r,n)}return r}function ie(e){for(var t=1;t<arguments.length;t++){var r=arguments[t]!=null?arguments[t]:{};t%2?oe(Object(r),!0).forEach(function(n){be(e,n,r[n])}):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(r)):oe(Object(r)).forEach(function(n){Object.defineProperty(e,n,Object.getOwnPropertyDescriptor(r,n))})}return e}function we(e,t){if(e==null)return{};var r={},n=Object.keys(e),o,i;for(i=0;i<n.length;i++)o=n[i],!(t.indexOf(o)>=0)&&(r[o]=e[o]);return r}function Oe(e,t){if(e==null)return{};var r=we(e,t),n,o;if(Object.getOwnPropertySymbols){var i=Object.getOwnPropertySymbols(e);for(o=0;o<i.length;o++)n=i[o],!(t.indexOf(n)>=0)&&Object.prototype.propertyIsEnumerable.call(e,n)&&(r[n]=e[n])}return r}function je(e,t){return Ee(e)||Me(e,t)||Se(e,t)||Pe()}function Ee(e){if(Array.isArray(e))return e}function Me(e,t){if(!(typeof Symbol>"u"||!(Symbol.iterator in Object(e)))){var r=[],n=!0,o=!1,i=void 0;try{for(var s=e[Symbol.iterator](),h;!(n=(h=s.next()).done)&&(r.push(h.value),!(t&&r.length===t));n=!0);}catch(p){o=!0,i=p}finally{try{!n&&s.return!=null&&s.return()}finally{if(o)throw i}}return r}}function Se(e,t){if(e){if(typeof e=="string")return ae(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);if(r==="Object"&&e.constructor&&(r=e.constructor.name),r==="Map"||r==="Set")return Array.from(e);if(r==="Arguments"||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return ae(e,t)}}function ae(e,t){(t==null||t>e.length)&&(t=e.length);for(var r=0,n=new Array(t);r<t;r++)n[r]=e[r];return n}function Pe(){throw new TypeError(`Invalid attempt to destructure non-iterable instance.
-In order to be iterable, non-array objects must have a [Symbol.iterator]() method.`)}function Re(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}function ue(e,t){var r=Object.keys(e);if(Object.getOwnPropertySymbols){var n=Object.getOwnPropertySymbols(e);t&&(n=n.filter(function(o){return Object.getOwnPropertyDescriptor(e,o).enumerable})),r.push.apply(r,n)}return r}function ce(e){for(var t=1;t<arguments.length;t++){var r=arguments[t]!=null?arguments[t]:{};t%2?ue(Object(r),!0).forEach(function(n){Re(e,n,r[n])}):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(r)):ue(Object(r)).forEach(function(n){Object.defineProperty(e,n,Object.getOwnPropertyDescriptor(r,n))})}return e}function Ie(){for(var e=arguments.length,t=new Array(e),r=0;r<e;r++)t[r]=arguments[r];return function(n){return t.reduceRight(function(o,i){return i(o)},n)}}function N(e){return function t(){for(var r=this,n=arguments.length,o=new Array(n),i=0;i<n;i++)o[i]=arguments[i];return o.length>=e.length?e.apply(this,o):function(){for(var s=arguments.length,h=new Array(s),p=0;p<s;p++)h[p]=arguments[p];return t.apply(r,[].concat(o,h))}}}function K(e){return{}.toString.call(e).includes("Object")}function Ae(e){return!Object.keys(e).length}function z(e){return typeof e=="function"}function Te(e,t){return Object.prototype.hasOwnProperty.call(e,t)}function Le(e,t){return K(t)||S("changeType"),Object.keys(t).some(function(r){return!Te(e,r)})&&S("changeField"),t}function Ce(e){z(e)||S("selectorType")}function $e(e){z(e)||K(e)||S("handlerType"),K(e)&&Object.values(e).some(function(t){return!z(t)})&&S("handlersType")}function xe(e){e||S("initialIsRequired"),K(e)||S("initialType"),Ae(e)&&S("initialContent")}function De(e,t){throw new Error(e[t]||e.default)}var Ve={initialIsRequired:"initial state is required",initialType:"initial state should be an object",initialContent:"initial state shouldn't be an empty object",handlerType:"handler should be an object or a function",handlersType:"all handlers should be a functions",selectorType:"selector should be a function",changeType:"provided value of changes should be an object",changeField:'it seams you want to change a field in the state which is not specified in the "initial" state',default:"an unknown error accured in `state-local` package"},S=N(De)(Ve),B={changes:Le,selector:Ce,handler:$e,initial:xe};function Ne(e){var t=arguments.length>1&&arguments[1]!==void 0?arguments[1]:{};B.initial(e),B.handler(t);var r={current:e},n=N(Ue)(r,t),o=N(ze)(r),i=N(B.changes)(e),s=N(qe)(r);function h(){var y=arguments.length>0&&arguments[0]!==void 0?arguments[0]:function(R){return R};return B.selector(y),y(r.current)}function p(y){Ie(n,o,i,s)(y)}return[h,p]}function qe(e,t){return z(t)?t(e.current):t}function ze(e,t){return e.current=ce(ce({},e.current),t),t}function Ue(e,t,r){return z(t)?t(e.current):Object.keys(r).forEach(function(n){var o;return(o=t[n])===null||o===void 0?void 0:o.call(t,e.current[n])}),r}var Fe={create:Ne},We={paths:{vs:"https://cdn.jsdelivr.net/npm/monaco-editor@0.36.1/min/vs"}};function He(e){return function t(){for(var r=this,n=arguments.length,o=new Array(n),i=0;i<n;i++)o[i]=arguments[i];return o.length>=e.length?e.apply(this,o):function(){for(var s=arguments.length,h=new Array(s),p=0;p<s;p++)h[p]=arguments[p];return t.apply(r,[].concat(o,h))}}}function Be(e){return{}.toString.call(e).includes("Object")}function Ge(e){return e||le("configIsRequired"),Be(e)||le("configType"),e.urls?(Ke(),{paths:{vs:e.urls.monacoBase}}):e}function Ke(){console.warn(se.deprecation)}function Ye(e,t){throw new Error(e[t]||e.default)}var se={configIsRequired:"the configuration object is required",configType:"the configuration object should be an object",default:"an unknown error accured in `@monaco-editor/loader` package",deprecation:`Deprecation warning!
-    You are using deprecated way of configuration.
-
-    Instead of using
-      monaco.config({ urls: { monacoBase: '...' } })
-    use
-      monaco.config({ paths: { vs: '...' } })
-
-    For more please check the link https://github.com/suren-atoyan/monaco-loader#config
-  `},le=He(Ye)(se),ke={config:Ge},Je=function(){for(var t=arguments.length,r=new Array(t),n=0;n<t;n++)r[n]=arguments[n];return function(o){return r.reduceRight(function(i,s){return s(i)},o)}};function fe(e,t){return Object.keys(t).forEach(function(r){t[r]instanceof Object&&e[r]&&Object.assign(t[r],fe(e[r],t[r]))}),ie(ie({},e),t)}var Qe={type:"cancelation",msg:"operation is manually canceled"};function te(e){var t=!1,r=new Promise(function(n,o){e.then(function(i){return t?o(Qe):n(i)}),e.catch(o)});return r.cancel=function(){return t=!0},r}var Xe=Fe.create({config:We,isInitialized:!1,resolve:null,reject:null,monaco:null}),de=je(Xe,2),U=de[0],k=de[1];function Ze(e){var t=ke.config(e),r=t.monaco,n=Oe(t,["monaco"]);k(function(o){return{config:fe(o.config,n),monaco:r}})}function _e(){var e=U(function(t){var r=t.monaco,n=t.isInitialized,o=t.resolve;return{monaco:r,isInitialized:n,resolve:o}});if(!e.isInitialized){if(k({isInitialized:!0}),e.monaco)return e.resolve(e.monaco),te(re);if(window.monaco&&window.monaco.editor)return pe(window.monaco),e.resolve(window.monaco),te(re);Je(et,rt)(nt)}return te(re)}function et(e){return document.body.appendChild(e)}function tt(e){var t=document.createElement("script");return e&&(t.src=e),t}function rt(e){var t=U(function(n){var o=n.config,i=n.reject;return{config:o,reject:i}}),r=tt("".concat(t.config.paths.vs,"/loader.js"));return r.onload=function(){return e()},r.onerror=t.reject,r}function nt(){var e=U(function(r){var n=r.config,o=r.resolve,i=r.reject;return{config:n,resolve:o,reject:i}}),t=window.require;t.config(e.config),t(["vs/editor/editor.main"],function(r){pe(r),e.resolve(r)},function(r){e.reject(r)})}function pe(e){U().monaco||k({monaco:e})}function ot(){return U(function(e){var t=e.monaco;return t})}var re=new Promise(function(e,t){return k({resolve:e,reject:t})}),ge={config:Ze,init:_e,__getMonacoInstance:ot},it={wrapper:{display:"flex",position:"relative",textAlign:"initial"},fullWidth:{width:"100%"},hide:{display:"none"}},ne=it,at={container:{display:"flex",height:"100%",width:"100%",justifyContent:"center",alignItems:"center"}},ut=at;function ct({children:e}){return L.createElement("div",{style:ut.container},e)}var lt=ct,st=lt;function ft({width:e,height:t,isEditorReady:r,loading:n,_ref:o,className:i,wrapperProps:s}){return L.createElement("section",{style:{...ne.wrapper,width:e,height:t},...s},!r&&L.createElement(st,null,n),L.createElement("div",{ref:o,style:{...ne.fullWidth,...!r&&ne.hide},className:i}))}var dt=ft,he=a.memo(dt);function pt(e){a.useEffect(e,[])}var ve=pt;function gt(e,t,r=!0){let n=a.useRef(!0);a.useEffect(n.current||!r?()=>{n.current=!1}:e,t)}var j=gt;function q(){}function Y(e,t,r,n){return ht(e,n)||vt(e,t,r,n)}function ht(e,t){return e.editor.getModel(me(e,t))}function vt(e,t,r,n){return e.editor.createModel(t,r,n?me(e,n):void 0)}function me(e,t){return e.Uri.parse(t)}function mt({original:e,modified:t,language:r,originalLanguage:n,modifiedLanguage:o,originalModelPath:i,modifiedModelPath:s,keepCurrentOriginalModel:h=!1,keepCurrentModifiedModel:p=!1,theme:y="light",loading:R="Loading...",options:E={},height:J="100%",width:Q="100%",className:X,wrapperProps:Z={},beforeMount:_=q,onMount:ee=q}){let[w,C]=a.useState(!1),[P,g]=a.useState(!0),O=a.useRef(null),v=a.useRef(null),$=a.useRef(null),m=a.useRef(ee),l=a.useRef(_),I=a.useRef(!1);ve(()=>{let u=ge.init();return u.then(d=>(v.current=d)&&g(!1)).catch(d=>(d==null?void 0:d.type)!=="cancelation"&&console.error("Monaco initialization: error:",d)),()=>O.current?x():u.cancel()}),j(()=>{let u=O.current.getModifiedEditor();u.getOption(v.current.editor.EditorOption.readOnly)?u.setValue(t||""):t!==u.getValue()&&(u.executeEdits("",[{range:u.getModel().getFullModelRange(),text:t||"",forceMoveMarkers:!0}]),u.pushUndoStop())},[t],w),j(()=>{var u,d;(d=(u=O.current)==null?void 0:u.getModel())==null||d.original.setValue(e||"")},[e],w),j(()=>{let{original:u,modified:d}=O.current.getModel();v.current.editor.setModelLanguage(u,n||r||"text"),v.current.editor.setModelLanguage(d,o||r||"text")},[r,n,o],w),j(()=>{var u;(u=v.current)==null||u.editor.setTheme(y)},[y],w),j(()=>{var u;(u=O.current)==null||u.updateOptions(E)},[E],w);let F=a.useCallback(()=>{var M;if(!v.current)return;l.current(v.current);let u=Y(v.current,e||"",n||r||"text",i||""),d=Y(v.current,t||"",o||r||"text",s||"");(M=O.current)==null||M.setModel({original:u,modified:d})},[r,t,o,e,n,i,s]),W=a.useCallback(()=>{var u;!I.current&&$.current&&(O.current=v.current.editor.createDiffEditor($.current,{automaticLayout:!0,...E}),F(),(u=v.current)==null||u.editor.setTheme(y),C(!0),I.current=!0)},[E,y,F]);a.useEffect(()=>{w&&m.current(O.current,v.current)},[w]),a.useEffect(()=>{!P&&!w&&W()},[P,w,W]);function x(){var d,M,A,D;let u=(d=O.current)==null?void 0:d.getModel();h||((M=u==null?void 0:u.original)==null||M.dispose()),p||((A=u==null?void 0:u.modified)==null||A.dispose()),(D=O.current)==null||D.dispose()}return L.createElement(he,{width:Q,height:J,isEditorReady:w,loading:R,_ref:$,className:X,wrapperProps:Z})}var yt=mt;a.memo(yt);function bt(e){let t=a.useRef();return a.useEffect(()=>{t.current=e},[e]),t.current}var wt=bt,G=new Map;function Ot({defaultValue:e,defaultLanguage:t,defaultPath:r,value:n,language:o,path:i,theme:s="light",line:h,loading:p="Loading...",options:y={},overrideServices:R={},saveViewState:E=!0,keepCurrentModel:J=!1,width:Q="100%",height:X="100%",className:Z,wrapperProps:_={},beforeMount:ee=q,onMount:w=q,onChange:C,onValidate:P=q}){let[g,O]=a.useState(!1),[v,$]=a.useState(!0),m=a.useRef(null),l=a.useRef(null),I=a.useRef(null),F=a.useRef(w),W=a.useRef(ee),x=a.useRef(),u=a.useRef(n),d=wt(i),M=a.useRef(!1),A=a.useRef(!1);ve(()=>{let c=ge.init();return c.then(f=>(m.current=f)&&$(!1)).catch(f=>(f==null?void 0:f.type)!=="cancelation"&&console.error("Monaco initialization: error:",f)),()=>l.current?ye():c.cancel()}),j(()=>{var f,b,V,T;let c=Y(m.current,e||n||"",t||o||"",i||r||"");c!==((f=l.current)==null?void 0:f.getModel())&&(E&&G.set(d,(b=l.current)==null?void 0:b.saveViewState()),(V=l.current)==null||V.setModel(c),E&&((T=l.current)==null||T.restoreViewState(G.get(i))))},[i],g),j(()=>{var c;(c=l.current)==null||c.updateOptions(y)},[y],g),j(()=>{!l.current||n===void 0||(l.current.getOption(m.current.editor.EditorOption.readOnly)?l.current.setValue(n):n!==l.current.getValue()&&(A.current=!0,l.current.executeEdits("",[{range:l.current.getModel().getFullModelRange(),text:n,forceMoveMarkers:!0}]),l.current.pushUndoStop(),A.current=!1))},[n],g),j(()=>{var f,b;let c=(f=l.current)==null?void 0:f.getModel();c&&o&&((b=m.current)==null||b.editor.setModelLanguage(c,o))},[o],g),j(()=>{var c;h!==void 0&&((c=l.current)==null||c.revealLine(h))},[h],g),j(()=>{var c;(c=m.current)==null||c.editor.setTheme(s)},[s],g);let D=a.useCallback(()=>{var c;if(!(!I.current||!m.current)&&!M.current){W.current(m.current);let f=i||r,b=Y(m.current,n||e||"",t||o||"",f||"");l.current=(c=m.current)==null?void 0:c.editor.create(I.current,{model:b,automaticLayout:!0,...y},R),E&&l.current.restoreViewState(G.get(f)),m.current.editor.setTheme(s),O(!0),M.current=!0}},[e,t,r,n,o,i,y,R,E,s]);a.useEffect(()=>{g&&F.current(l.current,m.current)},[g]),a.useEffect(()=>{!v&&!g&&D()},[v,g,D]),u.current=n,a.useEffect(()=>{var c,f;g&&C&&((c=x.current)==null||c.dispose(),x.current=(f=l.current)==null?void 0:f.onDidChangeModelContent(b=>{A.current||C(l.current.getValue(),b)}))},[g,C]),a.useEffect(()=>{if(g){let c=m.current.editor.onDidChangeMarkers(f=>{var V;let b=(V=l.current.getModel())==null?void 0:V.uri;if(b&&f.find(T=>T.path===b.path)){let T=m.current.editor.getModelMarkers({resource:b});P==null||P(T)}});return()=>{c==null||c.dispose()}}return()=>{}},[g,P]);function ye(){var c,f;(c=x.current)==null||c.dispose(),J?E&&G.set(i,l.current.saveViewState()):(f=l.current.getModel())==null||f.dispose(),l.current.dispose()}return L.createElement(he,{width:Q,height:X,isEditorReady:g,loading:p,_ref:I,className:Z,wrapperProps:_})}var jt=Ot,Et=a.memo(jt),Mt=Et;const St=()=>{const e={theme:"vs-dark"};return H.jsxs("div",{children:[H.jsx("h1",{children:"Playground"}),H.jsx("div",{style:{height:"50vh",width:"100%"},children:H.jsx(Mt,{...e})})]})};St.displayName="Top";export{St as default};
