--- docs/assets/Top-43ee8673.js
@@ -1,1 +1,0 @@
-import{j as e}from"./index-10388f2a.js";const o=()=>e.jsx("div",{children:e.jsx("h1",{children:"Top Page"})});o.displayName="Top";export{o as default};
