+++ docs/assets/Top-d2d68f9c.js
@@ -1,0 +1,1 @@
+import{j as e}from"./index-8bd2be33.js";const o=()=>e.jsx("div",{children:e.jsx("h1",{children:"Top Page"})});o.displayName="Top";export{o as default};
