+++ docs/assets/index-8bd2be33.js
@@ -1,0 +1,67 @@
+function Bc(e,t){for(var n=0;n<t.length;n++){const r=t[n];if(typeof r!="string"&&!Array.isArray(r)){for(const l in r)if(l!=="default"&&!(l in e)){const o=Object.getOwnPropertyDescriptor(r,l);o&&Object.defineProperty(e,l,o.get?o:{enumerable:!0,get:()=>r[l]})}}}return Object.freeze(Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}))}(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const l of document.querySelectorAll('link[rel="modulepreload"]'))r(l);new MutationObserver(l=>{for(const o of l)if(o.type==="childList")for(const i of o.addedNodes)i.tagName==="LINK"&&i.rel==="modulepreload"&&r(i)}).observe(document,{childList:!0,subtree:!0});function n(l){const o={};return l.integrity&&(o.integrity=l.integrity),l.referrerPolicy&&(o.referrerPolicy=l.referrerPolicy),l.crossOrigin==="use-credentials"?o.credentials="include":l.crossOrigin==="anonymous"?o.credentials="omit":o.credentials="same-origin",o}function r(l){if(l.ep)return;l.ep=!0;const o=n(l);fetch(l.href,o)}})();function Vc(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var K={},Wc={get exports(){return K},set exports(e){K=e}},sl={},C={},Ac={get exports(){return C},set exports(e){C=e}},T={};/**
+ * @license React
+ * react.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var er=Symbol.for("react.element"),Hc=Symbol.for("react.portal"),Qc=Symbol.for("react.fragment"),Kc=Symbol.for("react.strict_mode"),Yc=Symbol.for("react.profiler"),Xc=Symbol.for("react.provider"),Gc=Symbol.for("react.context"),Zc=Symbol.for("react.forward_ref"),Jc=Symbol.for("react.suspense"),qc=Symbol.for("react.memo"),bc=Symbol.for("react.lazy"),tu=Symbol.iterator;function ef(e){return e===null||typeof e!="object"?null:(e=tu&&e[tu]||e["@@iterator"],typeof e=="function"?e:null)}var ws={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},Ss=Object.assign,ks={};function sn(e,t,n){this.props=e,this.context=t,this.refs=ks,this.updater=n||ws}sn.prototype.isReactComponent={};sn.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")};sn.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function Es(){}Es.prototype=sn.prototype;function ri(e,t,n){this.props=e,this.context=t,this.refs=ks,this.updater=n||ws}var li=ri.prototype=new Es;li.constructor=ri;Ss(li,sn.prototype);li.isPureReactComponent=!0;var nu=Array.isArray,xs=Object.prototype.hasOwnProperty,oi={current:null},Cs={key:!0,ref:!0,__self:!0,__source:!0};function _s(e,t,n){var r,l={},o=null,i=null;if(t!=null)for(r in t.ref!==void 0&&(i=t.ref),t.key!==void 0&&(o=""+t.key),t)xs.call(t,r)&&!Cs.hasOwnProperty(r)&&(l[r]=t[r]);var u=arguments.length-2;if(u===1)l.children=n;else if(1<u){for(var s=Array(u),a=0;a<u;a++)s[a]=arguments[a+2];l.children=s}if(e&&e.defaultProps)for(r in u=e.defaultProps,u)l[r]===void 0&&(l[r]=u[r]);return{$$typeof:er,type:e,key:o,ref:i,props:l,_owner:oi.current}}function tf(e,t){return{$$typeof:er,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function ii(e){return typeof e=="object"&&e!==null&&e.$$typeof===er}function nf(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var ru=/\/+/g;function zl(e,t){return typeof e=="object"&&e!==null&&e.key!=null?nf(""+e.key):t.toString(36)}function Cr(e,t,n,r,l){var o=typeof e;(o==="undefined"||o==="boolean")&&(e=null);var i=!1;if(e===null)i=!0;else switch(o){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case er:case Hc:i=!0}}if(i)return i=e,l=l(i),e=r===""?"."+zl(i,0):r,nu(l)?(n="",e!=null&&(n=e.replace(ru,"$&/")+"/"),Cr(l,t,n,"",function(a){return a})):l!=null&&(ii(l)&&(l=tf(l,n+(!l.key||i&&i.key===l.key?"":(""+l.key).replace(ru,"$&/")+"/")+e)),t.push(l)),1;if(i=0,r=r===""?".":r+":",nu(e))for(var u=0;u<e.length;u++){o=e[u];var s=r+zl(o,u);i+=Cr(o,t,n,s,l)}else if(s=ef(e),typeof s=="function")for(e=s.call(e),u=0;!(o=e.next()).done;)o=o.value,s=r+zl(o,u++),i+=Cr(o,t,n,s,l);else if(o==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return i}function ur(e,t,n){if(e==null)return e;var r=[],l=0;return Cr(e,r,"","",function(o){return t.call(n,o,l++)}),r}function rf(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var ae={current:null},_r={transition:null},lf={ReactCurrentDispatcher:ae,ReactCurrentBatchConfig:_r,ReactCurrentOwner:oi};T.Children={map:ur,forEach:function(e,t,n){ur(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return ur(e,function(){t++}),t},toArray:function(e){return ur(e,function(t){return t})||[]},only:function(e){if(!ii(e))throw Error("React.Children.only expected to receive a single React element child.");return e}};T.Component=sn;T.Fragment=Qc;T.Profiler=Yc;T.PureComponent=ri;T.StrictMode=Kc;T.Suspense=Jc;T.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=lf;T.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=Ss({},e.props),l=e.key,o=e.ref,i=e._owner;if(t!=null){if(t.ref!==void 0&&(o=t.ref,i=oi.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)xs.call(t,s)&&!Cs.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&u!==void 0?u[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){u=Array(s);for(var a=0;a<s;a++)u[a]=arguments[a+2];r.children=u}return{$$typeof:er,type:e.type,key:l,ref:o,props:r,_owner:i}};T.createContext=function(e){return e={$$typeof:Gc,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:Xc,_context:e},e.Consumer=e};T.createElement=_s;T.createFactory=function(e){var t=_s.bind(null,e);return t.type=e,t};T.createRef=function(){return{current:null}};T.forwardRef=function(e){return{$$typeof:Zc,render:e}};T.isValidElement=ii;T.lazy=function(e){return{$$typeof:bc,_payload:{_status:-1,_result:e},_init:rf}};T.memo=function(e,t){return{$$typeof:qc,type:e,compare:t===void 0?null:t}};T.startTransition=function(e){var t=_r.transition;_r.transition={};try{e()}finally{_r.transition=t}};T.unstable_act=function(){throw Error("act(...) is not supported in production builds of React.")};T.useCallback=function(e,t){return ae.current.useCallback(e,t)};T.useContext=function(e){return ae.current.useContext(e)};T.useDebugValue=function(){};T.useDeferredValue=function(e){return ae.current.useDeferredValue(e)};T.useEffect=function(e,t){return ae.current.useEffect(e,t)};T.useId=function(){return ae.current.useId()};T.useImperativeHandle=function(e,t,n){return ae.current.useImperativeHandle(e,t,n)};T.useInsertionEffect=function(e,t){return ae.current.useInsertionEffect(e,t)};T.useLayoutEffect=function(e,t){return ae.current.useLayoutEffect(e,t)};T.useMemo=function(e,t){return ae.current.useMemo(e,t)};T.useReducer=function(e,t,n){return ae.current.useReducer(e,t,n)};T.useRef=function(e){return ae.current.useRef(e)};T.useState=function(e){return ae.current.useState(e)};T.useSyncExternalStore=function(e,t,n){return ae.current.useSyncExternalStore(e,t,n)};T.useTransition=function(){return ae.current.useTransition()};T.version="18.2.0";(function(e){e.exports=T})(Ac);const Ps=Vc(C),ro=Bc({__proto__:null,default:Ps},[C]);/**
+ * @license React
+ * react-jsx-runtime.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var of=C,uf=Symbol.for("react.element"),sf=Symbol.for("react.fragment"),af=Object.prototype.hasOwnProperty,cf=of.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,ff={key:!0,ref:!0,__self:!0,__source:!0};function Ns(e,t,n){var r,l={},o=null,i=null;n!==void 0&&(o=""+n),t.key!==void 0&&(o=""+t.key),t.ref!==void 0&&(i=t.ref);for(r in t)af.call(t,r)&&!ff.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:uf,type:e,key:o,ref:i,props:l,_owner:cf.current}}sl.Fragment=sf;sl.jsx=Ns;sl.jsxs=Ns;(function(e){e.exports=sl})(Wc);var lo={},oo={},df={get exports(){return oo},set exports(e){oo=e}},Se={},io={},pf={get exports(){return io},set exports(e){io=e}},Ls={};/**
+ * @license React
+ * scheduler.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */(function(e){function t(x,z){var R=x.length;x.push(z);e:for(;0<R;){var H=R-1>>>1,Z=x[H];if(0<l(Z,z))x[H]=z,x[R]=Z,R=H;else break e}}function n(x){return x.length===0?null:x[0]}function r(x){if(x.length===0)return null;var z=x[0],R=x.pop();if(R!==z){x[0]=R;e:for(var H=0,Z=x.length,or=Z>>>1;H<or;){var gt=2*(H+1)-1,Ll=x[gt],wt=gt+1,ir=x[wt];if(0>l(Ll,R))wt<Z&&0>l(ir,Ll)?(x[H]=ir,x[wt]=R,H=wt):(x[H]=Ll,x[gt]=R,H=gt);else if(wt<Z&&0>l(ir,R))x[H]=ir,x[wt]=R,H=wt;else break e}}return z}function l(x,z){var R=x.sortIndex-z.sortIndex;return R!==0?R:x.id-z.id}if(typeof performance=="object"&&typeof performance.now=="function"){var o=performance;e.unstable_now=function(){return o.now()}}else{var i=Date,u=i.now();e.unstable_now=function(){return i.now()-u}}var s=[],a=[],h=1,m=null,p=3,w=!1,S=!1,g=!1,L=typeof setTimeout=="function"?setTimeout:null,f=typeof clearTimeout=="function"?clearTimeout:null,c=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function d(x){for(var z=n(a);z!==null;){if(z.callback===null)r(a);else if(z.startTime<=x)r(a),z.sortIndex=z.expirationTime,t(s,z);else break;z=n(a)}}function v(x){if(g=!1,d(x),!S)if(n(s)!==null)S=!0,Pl(E);else{var z=n(a);z!==null&&Nl(v,z.startTime-x)}}function E(x,z){S=!1,g&&(g=!1,f(N),N=-1),w=!0;var R=p;try{for(d(z),m=n(s);m!==null&&(!(m.expirationTime>z)||x&&!Le());){var H=m.callback;if(typeof H=="function"){m.callback=null,p=m.priorityLevel;var Z=H(m.expirationTime<=z);z=e.unstable_now(),typeof Z=="function"?m.callback=Z:m===n(s)&&r(s),d(z)}else r(s);m=n(s)}if(m!==null)var or=!0;else{var gt=n(a);gt!==null&&Nl(v,gt.startTime-z),or=!1}return or}finally{m=null,p=R,w=!1}}var _=!1,P=null,N=-1,A=5,O=-1;function Le(){return!(e.unstable_now()-O<A)}function pn(){if(P!==null){var x=e.unstable_now();O=x;var z=!0;try{z=P(!0,x)}finally{z?hn():(_=!1,P=null)}}else _=!1}var hn;if(typeof c=="function")hn=function(){c(pn)};else if(typeof MessageChannel<"u"){var eu=new MessageChannel,$c=eu.port2;eu.port1.onmessage=pn,hn=function(){$c.postMessage(null)}}else hn=function(){L(pn,0)};function Pl(x){P=x,_||(_=!0,hn())}function Nl(x,z){N=L(function(){x(e.unstable_now())},z)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(x){x.callback=null},e.unstable_continueExecution=function(){S||w||(S=!0,Pl(E))},e.unstable_forceFrameRate=function(x){0>x||125<x?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):A=0<x?Math.floor(1e3/x):5},e.unstable_getCurrentPriorityLevel=function(){return p},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(x){switch(p){case 1:case 2:case 3:var z=3;break;default:z=p}var R=p;p=z;try{return x()}finally{p=R}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(x,z){switch(x){case 1:case 2:case 3:case 4:case 5:break;default:x=3}var R=p;p=x;try{return z()}finally{p=R}},e.unstable_scheduleCallback=function(x,z,R){var H=e.unstable_now();switch(typeof R=="object"&&R!==null?(R=R.delay,R=typeof R=="number"&&0<R?H+R:H):R=H,x){case 1:var Z=-1;break;case 2:Z=250;break;case 5:Z=1073741823;break;case 4:Z=1e4;break;default:Z=5e3}return Z=R+Z,x={id:h++,callback:z,priorityLevel:x,startTime:R,expirationTime:Z,sortIndex:-1},R>H?(x.sortIndex=R,t(a,x),n(s)===null&&x===n(a)&&(g?(f(N),N=-1):g=!0,Nl(v,R-H))):(x.sortIndex=Z,t(s,x),S||w||(S=!0,Pl(E))),x},e.unstable_shouldYield=Le,e.unstable_wrapCallback=function(x){var z=p;return function(){var R=p;p=z;try{return x.apply(this,arguments)}finally{p=R}}}})(Ls);(function(e){e.exports=Ls})(pf);/**
+ * @license React
+ * react-dom.production.min.js
+ *
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */var zs=C,we=io;function y(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var Rs=new Set,Dn={};function It(e,t){en(e,t),en(e+"Capture",t)}function en(e,t){for(Dn[e]=t,e=0;e<t.length;e++)Rs.add(t[e])}var Ke=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),uo=Object.prototype.hasOwnProperty,hf=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,lu={},ou={};function mf(e){return uo.call(ou,e)?!0:uo.call(lu,e)?!1:hf.test(e)?ou[e]=!0:(lu[e]=!0,!1)}function vf(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function yf(e,t,n,r){if(t===null||typeof t>"u"||vf(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function ce(e,t,n,r,l,o,i){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=o,this.removeEmptyString=i}var ne={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){ne[e]=new ce(e,0,!1,e,null,!1,!1)});[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];ne[t]=new ce(t,1,!1,e[1],null,!1,!1)});["contentEditable","draggable","spellCheck","value"].forEach(function(e){ne[e]=new ce(e,2,!1,e.toLowerCase(),null,!1,!1)});["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){ne[e]=new ce(e,2,!1,e,null,!1,!1)});"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){ne[e]=new ce(e,3,!1,e.toLowerCase(),null,!1,!1)});["checked","multiple","muted","selected"].forEach(function(e){ne[e]=new ce(e,3,!0,e,null,!1,!1)});["capture","download"].forEach(function(e){ne[e]=new ce(e,4,!1,e,null,!1,!1)});["cols","rows","size","span"].forEach(function(e){ne[e]=new ce(e,6,!1,e,null,!1,!1)});["rowSpan","start"].forEach(function(e){ne[e]=new ce(e,5,!1,e.toLowerCase(),null,!1,!1)});var ui=/[\-:]([a-z])/g;function si(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(ui,si);ne[t]=new ce(t,1,!1,e,null,!1,!1)});"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(ui,si);ne[t]=new ce(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)});["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(ui,si);ne[t]=new ce(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)});["tabIndex","crossOrigin"].forEach(function(e){ne[e]=new ce(e,1,!1,e.toLowerCase(),null,!1,!1)});ne.xlinkHref=new ce("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1);["src","href","action","formAction"].forEach(function(e){ne[e]=new ce(e,1,!1,e.toLowerCase(),null,!0,!0)});function ai(e,t,n,r){var l=ne.hasOwnProperty(t)?ne[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(yf(t,n,l,r)&&(n=null),r||l===null?mf(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var Ze=zs.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,sr=Symbol.for("react.element"),Dt=Symbol.for("react.portal"),Ft=Symbol.for("react.fragment"),ci=Symbol.for("react.strict_mode"),so=Symbol.for("react.profiler"),Ts=Symbol.for("react.provider"),Os=Symbol.for("react.context"),fi=Symbol.for("react.forward_ref"),ao=Symbol.for("react.suspense"),co=Symbol.for("react.suspense_list"),di=Symbol.for("react.memo"),qe=Symbol.for("react.lazy"),Is=Symbol.for("react.offscreen"),iu=Symbol.iterator;function mn(e){return e===null||typeof e!="object"?null:(e=iu&&e[iu]||e["@@iterator"],typeof e=="function"?e:null)}var V=Object.assign,Rl;function xn(e){if(Rl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Rl=t&&t[1]||""}return`
+`+Rl+e}var Tl=!1;function Ol(e,t){if(!e||Tl)return"";Tl=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(a){var r=a}Reflect.construct(e,[],t)}else{try{t.call()}catch(a){r=a}e.call(t.prototype)}else{try{throw Error()}catch(a){r=a}e()}}catch(a){if(a&&r&&typeof a.stack=="string"){for(var l=a.stack.split(`
+`),o=r.stack.split(`
+`),i=l.length-1,u=o.length-1;1<=i&&0<=u&&l[i]!==o[u];)u--;for(;1<=i&&0<=u;i--,u--)if(l[i]!==o[u]){if(i!==1||u!==1)do if(i--,u--,0>u||l[i]!==o[u]){var s=`
+`+l[i].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=i&&0<=u);break}}}finally{Tl=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?xn(e):""}function gf(e){switch(e.tag){case 5:return xn(e.type);case 16:return xn("Lazy");case 13:return xn("Suspense");case 19:return xn("SuspenseList");case 0:case 2:case 15:return e=Ol(e.type,!1),e;case 11:return e=Ol(e.type.render,!1),e;case 1:return e=Ol(e.type,!0),e;default:return""}}function fo(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case Ft:return"Fragment";case Dt:return"Portal";case so:return"Profiler";case ci:return"StrictMode";case ao:return"Suspense";case co:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case Os:return(e.displayName||"Context")+".Consumer";case Ts:return(e._context.displayName||"Context")+".Provider";case fi:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case di:return t=e.displayName||null,t!==null?t:fo(e.type)||"Memo";case qe:t=e._payload,e=e._init;try{return fo(e(t))}catch{}}return null}function wf(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return fo(t);case 8:return t===ci?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function pt(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function js(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function Sf(e){var t=js(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,o=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(i){r=""+i,o.call(this,i)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(i){r=""+i},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function ar(e){e._valueTracker||(e._valueTracker=Sf(e))}function Ms(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=js(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function Fr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function po(e,t){var n=t.checked;return V({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function uu(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=pt(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function Ds(e,t){t=t.checked,t!=null&&ai(e,"checked",t,!1)}function ho(e,t){Ds(e,t);var n=pt(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?mo(e,t.type,n):t.hasOwnProperty("defaultValue")&&mo(e,t.type,pt(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function su(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function mo(e,t,n){(t!=="number"||Fr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var Cn=Array.isArray;function Xt(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+pt(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function vo(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(y(91));return V({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function au(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(y(92));if(Cn(n)){if(1<n.length)throw Error(y(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:pt(n)}}function Fs(e,t){var n=pt(t.value),r=pt(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function cu(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function Us(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function yo(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?Us(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var cr,$s=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(cr=cr||document.createElement("div"),cr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=cr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function Fn(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var Nn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},kf=["Webkit","ms","Moz","O"];Object.keys(Nn).forEach(function(e){kf.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),Nn[t]=Nn[e]})});function Bs(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||Nn.hasOwnProperty(e)&&Nn[e]?(""+t).trim():t+"px"}function Vs(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=Bs(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var Ef=V({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function go(e,t){if(t){if(Ef[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(y(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(y(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(y(61))}if(t.style!=null&&typeof t.style!="object")throw Error(y(62))}}function wo(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var So=null;function pi(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var ko=null,Gt=null,Zt=null;function fu(e){if(e=rr(e)){if(typeof ko!="function")throw Error(y(280));var t=e.stateNode;t&&(t=pl(t),ko(e.stateNode,e.type,t))}}function Ws(e){Gt?Zt?Zt.push(e):Zt=[e]:Gt=e}function As(){if(Gt){var e=Gt,t=Zt;if(Zt=Gt=null,fu(e),t)for(e=0;e<t.length;e++)fu(t[e])}}function Hs(e,t){return e(t)}function Qs(){}var Il=!1;function Ks(e,t,n){if(Il)return e(t,n);Il=!0;try{return Hs(e,t,n)}finally{Il=!1,(Gt!==null||Zt!==null)&&(Qs(),As())}}function Un(e,t){var n=e.stateNode;if(n===null)return null;var r=pl(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(y(231,t,typeof n));return n}var Eo=!1;if(Ke)try{var vn={};Object.defineProperty(vn,"passive",{get:function(){Eo=!0}}),window.addEventListener("test",vn,vn),window.removeEventListener("test",vn,vn)}catch{Eo=!1}function xf(e,t,n,r,l,o,i,u,s){var a=Array.prototype.slice.call(arguments,3);try{t.apply(n,a)}catch(h){this.onError(h)}}var Ln=!1,Ur=null,$r=!1,xo=null,Cf={onError:function(e){Ln=!0,Ur=e}};function _f(e,t,n,r,l,o,i,u,s){Ln=!1,Ur=null,xf.apply(Cf,arguments)}function Pf(e,t,n,r,l,o,i,u,s){if(_f.apply(this,arguments),Ln){if(Ln){var a=Ur;Ln=!1,Ur=null}else throw Error(y(198));$r||($r=!0,xo=a)}}function jt(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function Ys(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function du(e){if(jt(e)!==e)throw Error(y(188))}function Nf(e){var t=e.alternate;if(!t){if(t=jt(e),t===null)throw Error(y(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var o=l.alternate;if(o===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===o.child){for(o=l.child;o;){if(o===n)return du(l),e;if(o===r)return du(l),t;o=o.sibling}throw Error(y(188))}if(n.return!==r.return)n=l,r=o;else{for(var i=!1,u=l.child;u;){if(u===n){i=!0,n=l,r=o;break}if(u===r){i=!0,r=l,n=o;break}u=u.sibling}if(!i){for(u=o.child;u;){if(u===n){i=!0,n=o,r=l;break}if(u===r){i=!0,r=o,n=l;break}u=u.sibling}if(!i)throw Error(y(189))}}if(n.alternate!==r)throw Error(y(190))}if(n.tag!==3)throw Error(y(188));return n.stateNode.current===n?e:t}function Xs(e){return e=Nf(e),e!==null?Gs(e):null}function Gs(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=Gs(e);if(t!==null)return t;e=e.sibling}return null}var Zs=we.unstable_scheduleCallback,pu=we.unstable_cancelCallback,Lf=we.unstable_shouldYield,zf=we.unstable_requestPaint,Q=we.unstable_now,Rf=we.unstable_getCurrentPriorityLevel,hi=we.unstable_ImmediatePriority,Js=we.unstable_UserBlockingPriority,Br=we.unstable_NormalPriority,Tf=we.unstable_LowPriority,qs=we.unstable_IdlePriority,al=null,$e=null;function Of(e){if($e&&typeof $e.onCommitFiberRoot=="function")try{$e.onCommitFiberRoot(al,e,void 0,(e.current.flags&128)===128)}catch{}}var Ie=Math.clz32?Math.clz32:Mf,If=Math.log,jf=Math.LN2;function Mf(e){return e>>>=0,e===0?32:31-(If(e)/jf|0)|0}var fr=64,dr=4194304;function _n(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Vr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,o=e.pingedLanes,i=n&268435455;if(i!==0){var u=i&~l;u!==0?r=_n(u):(o&=i,o!==0&&(r=_n(o)))}else i=n&~l,i!==0?r=_n(i):o!==0&&(r=_n(o));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,o=t&-t,l>=o||l===16&&(o&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Ie(t),l=1<<n,r|=e[n],t&=~l;return r}function Df(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Ff(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,o=e.pendingLanes;0<o;){var i=31-Ie(o),u=1<<i,s=l[i];s===-1?(!(u&n)||u&r)&&(l[i]=Df(u,t)):s<=t&&(e.expiredLanes|=u),o&=~u}}function Co(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function bs(){var e=fr;return fr<<=1,!(fr&4194240)&&(fr=64),e}function jl(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function tr(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Ie(t),e[t]=n}function Uf(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Ie(n),o=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~o}}function mi(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Ie(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var j=0;function ea(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var ta,vi,na,ra,la,_o=!1,pr=[],ot=null,it=null,ut=null,$n=new Map,Bn=new Map,et=[],$f="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function hu(e,t){switch(e){case"focusin":case"focusout":ot=null;break;case"dragenter":case"dragleave":it=null;break;case"mouseover":case"mouseout":ut=null;break;case"pointerover":case"pointerout":$n.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Bn.delete(t.pointerId)}}function yn(e,t,n,r,l,o){return e===null||e.nativeEvent!==o?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:o,targetContainers:[l]},t!==null&&(t=rr(t),t!==null&&vi(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Bf(e,t,n,r,l){switch(t){case"focusin":return ot=yn(ot,e,t,n,r,l),!0;case"dragenter":return it=yn(it,e,t,n,r,l),!0;case"mouseover":return ut=yn(ut,e,t,n,r,l),!0;case"pointerover":var o=l.pointerId;return $n.set(o,yn($n.get(o)||null,e,t,n,r,l)),!0;case"gotpointercapture":return o=l.pointerId,Bn.set(o,yn(Bn.get(o)||null,e,t,n,r,l)),!0}return!1}function oa(e){var t=Et(e.target);if(t!==null){var n=jt(t);if(n!==null){if(t=n.tag,t===13){if(t=Ys(n),t!==null){e.blockedOn=t,la(e.priority,function(){na(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Pr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=Po(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);So=r,n.target.dispatchEvent(r),So=null}else return t=rr(n),t!==null&&vi(t),e.blockedOn=n,!1;t.shift()}return!0}function mu(e,t,n){Pr(e)&&n.delete(t)}function Vf(){_o=!1,ot!==null&&Pr(ot)&&(ot=null),it!==null&&Pr(it)&&(it=null),ut!==null&&Pr(ut)&&(ut=null),$n.forEach(mu),Bn.forEach(mu)}function gn(e,t){e.blockedOn===t&&(e.blockedOn=null,_o||(_o=!0,we.unstable_scheduleCallback(we.unstable_NormalPriority,Vf)))}function Vn(e){function t(l){return gn(l,e)}if(0<pr.length){gn(pr[0],e);for(var n=1;n<pr.length;n++){var r=pr[n];r.blockedOn===e&&(r.blockedOn=null)}}for(ot!==null&&gn(ot,e),it!==null&&gn(it,e),ut!==null&&gn(ut,e),$n.forEach(t),Bn.forEach(t),n=0;n<et.length;n++)r=et[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<et.length&&(n=et[0],n.blockedOn===null);)oa(n),n.blockedOn===null&&et.shift()}var Jt=Ze.ReactCurrentBatchConfig,Wr=!0;function Wf(e,t,n,r){var l=j,o=Jt.transition;Jt.transition=null;try{j=1,yi(e,t,n,r)}finally{j=l,Jt.transition=o}}function Af(e,t,n,r){var l=j,o=Jt.transition;Jt.transition=null;try{j=4,yi(e,t,n,r)}finally{j=l,Jt.transition=o}}function yi(e,t,n,r){if(Wr){var l=Po(e,t,n,r);if(l===null)Hl(e,t,r,Ar,n),hu(e,r);else if(Bf(l,e,t,n,r))r.stopPropagation();else if(hu(e,r),t&4&&-1<$f.indexOf(e)){for(;l!==null;){var o=rr(l);if(o!==null&&ta(o),o=Po(e,t,n,r),o===null&&Hl(e,t,r,Ar,n),o===l)break;l=o}l!==null&&r.stopPropagation()}else Hl(e,t,r,null,n)}}var Ar=null;function Po(e,t,n,r){if(Ar=null,e=pi(r),e=Et(e),e!==null)if(t=jt(e),t===null)e=null;else if(n=t.tag,n===13){if(e=Ys(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Ar=e,null}function ia(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Rf()){case hi:return 1;case Js:return 4;case Br:case Tf:return 16;case qs:return 536870912;default:return 16}default:return 16}}var nt=null,gi=null,Nr=null;function ua(){if(Nr)return Nr;var e,t=gi,n=t.length,r,l="value"in nt?nt.value:nt.textContent,o=l.length;for(e=0;e<n&&t[e]===l[e];e++);var i=n-e;for(r=1;r<=i&&t[n-r]===l[o-r];r++);return Nr=l.slice(e,1<r?1-r:void 0)}function Lr(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function hr(){return!0}function vu(){return!1}function ke(e){function t(n,r,l,o,i){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=o,this.target=i,this.currentTarget=null;for(var u in e)e.hasOwnProperty(u)&&(n=e[u],this[u]=n?n(o):o[u]);return this.isDefaultPrevented=(o.defaultPrevented!=null?o.defaultPrevented:o.returnValue===!1)?hr:vu,this.isPropagationStopped=vu,this}return V(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=hr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=hr)},persist:function(){},isPersistent:hr}),t}var an={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},wi=ke(an),nr=V({},an,{view:0,detail:0}),Hf=ke(nr),Ml,Dl,wn,cl=V({},nr,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Si,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==wn&&(wn&&e.type==="mousemove"?(Ml=e.screenX-wn.screenX,Dl=e.screenY-wn.screenY):Dl=Ml=0,wn=e),Ml)},movementY:function(e){return"movementY"in e?e.movementY:Dl}}),yu=ke(cl),Qf=V({},cl,{dataTransfer:0}),Kf=ke(Qf),Yf=V({},nr,{relatedTarget:0}),Fl=ke(Yf),Xf=V({},an,{animationName:0,elapsedTime:0,pseudoElement:0}),Gf=ke(Xf),Zf=V({},an,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Jf=ke(Zf),qf=V({},an,{data:0}),gu=ke(qf),bf={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},ed={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},td={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function nd(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=td[e])?!!t[e]:!1}function Si(){return nd}var rd=V({},nr,{key:function(e){if(e.key){var t=bf[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Lr(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?ed[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Si,charCode:function(e){return e.type==="keypress"?Lr(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Lr(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),ld=ke(rd),od=V({},cl,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),wu=ke(od),id=V({},nr,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Si}),ud=ke(id),sd=V({},an,{propertyName:0,elapsedTime:0,pseudoElement:0}),ad=ke(sd),cd=V({},cl,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),fd=ke(cd),dd=[9,13,27,32],ki=Ke&&"CompositionEvent"in window,zn=null;Ke&&"documentMode"in document&&(zn=document.documentMode);var pd=Ke&&"TextEvent"in window&&!zn,sa=Ke&&(!ki||zn&&8<zn&&11>=zn),Su=String.fromCharCode(32),ku=!1;function aa(e,t){switch(e){case"keyup":return dd.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function ca(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var Ut=!1;function hd(e,t){switch(e){case"compositionend":return ca(t);case"keypress":return t.which!==32?null:(ku=!0,Su);case"textInput":return e=t.data,e===Su&&ku?null:e;default:return null}}function md(e,t){if(Ut)return e==="compositionend"||!ki&&aa(e,t)?(e=ua(),Nr=gi=nt=null,Ut=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return sa&&t.locale!=="ko"?null:t.data;default:return null}}var vd={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Eu(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!vd[e.type]:t==="textarea"}function fa(e,t,n,r){Ws(r),t=Hr(t,"onChange"),0<t.length&&(n=new wi("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var Rn=null,Wn=null;function yd(e){Ea(e,0)}function fl(e){var t=Vt(e);if(Ms(t))return e}function gd(e,t){if(e==="change")return t}var da=!1;if(Ke){var Ul;if(Ke){var $l="oninput"in document;if(!$l){var xu=document.createElement("div");xu.setAttribute("oninput","return;"),$l=typeof xu.oninput=="function"}Ul=$l}else Ul=!1;da=Ul&&(!document.documentMode||9<document.documentMode)}function Cu(){Rn&&(Rn.detachEvent("onpropertychange",pa),Wn=Rn=null)}function pa(e){if(e.propertyName==="value"&&fl(Wn)){var t=[];fa(t,Wn,e,pi(e)),Ks(yd,t)}}function wd(e,t,n){e==="focusin"?(Cu(),Rn=t,Wn=n,Rn.attachEvent("onpropertychange",pa)):e==="focusout"&&Cu()}function Sd(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return fl(Wn)}function kd(e,t){if(e==="click")return fl(t)}function Ed(e,t){if(e==="input"||e==="change")return fl(t)}function xd(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Me=typeof Object.is=="function"?Object.is:xd;function An(e,t){if(Me(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!uo.call(t,l)||!Me(e[l],t[l]))return!1}return!0}function _u(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function Pu(e,t){var n=_u(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=_u(n)}}function ha(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?ha(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function ma(){for(var e=window,t=Fr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=Fr(e.document)}return t}function Ei(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function Cd(e){var t=ma(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&ha(n.ownerDocument.documentElement,n)){if(r!==null&&Ei(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,o=Math.min(r.start,l);r=r.end===void 0?o:Math.min(r.end,l),!e.extend&&o>r&&(l=r,r=o,o=l),l=Pu(n,o);var i=Pu(n,r);l&&i&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==i.node||e.focusOffset!==i.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),o>r?(e.addRange(t),e.extend(i.node,i.offset)):(t.setEnd(i.node,i.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var _d=Ke&&"documentMode"in document&&11>=document.documentMode,$t=null,No=null,Tn=null,Lo=!1;function Nu(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;Lo||$t==null||$t!==Fr(r)||(r=$t,"selectionStart"in r&&Ei(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Tn&&An(Tn,r)||(Tn=r,r=Hr(No,"onSelect"),0<r.length&&(t=new wi("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=$t)))}function mr(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var Bt={animationend:mr("Animation","AnimationEnd"),animationiteration:mr("Animation","AnimationIteration"),animationstart:mr("Animation","AnimationStart"),transitionend:mr("Transition","TransitionEnd")},Bl={},va={};Ke&&(va=document.createElement("div").style,"AnimationEvent"in window||(delete Bt.animationend.animation,delete Bt.animationiteration.animation,delete Bt.animationstart.animation),"TransitionEvent"in window||delete Bt.transitionend.transition);function dl(e){if(Bl[e])return Bl[e];if(!Bt[e])return e;var t=Bt[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in va)return Bl[e]=t[n];return e}var ya=dl("animationend"),ga=dl("animationiteration"),wa=dl("animationstart"),Sa=dl("transitionend"),ka=new Map,Lu="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function mt(e,t){ka.set(e,t),It(t,[e])}for(var Vl=0;Vl<Lu.length;Vl++){var Wl=Lu[Vl],Pd=Wl.toLowerCase(),Nd=Wl[0].toUpperCase()+Wl.slice(1);mt(Pd,"on"+Nd)}mt(ya,"onAnimationEnd");mt(ga,"onAnimationIteration");mt(wa,"onAnimationStart");mt("dblclick","onDoubleClick");mt("focusin","onFocus");mt("focusout","onBlur");mt(Sa,"onTransitionEnd");en("onMouseEnter",["mouseout","mouseover"]);en("onMouseLeave",["mouseout","mouseover"]);en("onPointerEnter",["pointerout","pointerover"]);en("onPointerLeave",["pointerout","pointerover"]);It("onChange","change click focusin focusout input keydown keyup selectionchange".split(" "));It("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" "));It("onBeforeInput",["compositionend","keypress","textInput","paste"]);It("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" "));It("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" "));It("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Pn="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Ld=new Set("cancel close invalid load scroll toggle".split(" ").concat(Pn));function zu(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,Pf(r,t,void 0,e),e.currentTarget=null}function Ea(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var o=void 0;if(t)for(var i=r.length-1;0<=i;i--){var u=r[i],s=u.instance,a=u.currentTarget;if(u=u.listener,s!==o&&l.isPropagationStopped())break e;zu(l,u,a),o=s}else for(i=0;i<r.length;i++){if(u=r[i],s=u.instance,a=u.currentTarget,u=u.listener,s!==o&&l.isPropagationStopped())break e;zu(l,u,a),o=s}}}if($r)throw e=xo,$r=!1,xo=null,e}function D(e,t){var n=t[Io];n===void 0&&(n=t[Io]=new Set);var r=e+"__bubble";n.has(r)||(xa(t,e,2,!1),n.add(r))}function Al(e,t,n){var r=0;t&&(r|=4),xa(n,e,r,t)}var vr="_reactListening"+Math.random().toString(36).slice(2);function Hn(e){if(!e[vr]){e[vr]=!0,Rs.forEach(function(n){n!=="selectionchange"&&(Ld.has(n)||Al(n,!1,e),Al(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[vr]||(t[vr]=!0,Al("selectionchange",!1,t))}}function xa(e,t,n,r){switch(ia(t)){case 1:var l=Wf;break;case 4:l=Af;break;default:l=yi}n=l.bind(null,t,n,e),l=void 0,!Eo||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Hl(e,t,n,r,l){var o=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var i=r.tag;if(i===3||i===4){var u=r.stateNode.containerInfo;if(u===l||u.nodeType===8&&u.parentNode===l)break;if(i===4)for(i=r.return;i!==null;){var s=i.tag;if((s===3||s===4)&&(s=i.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;i=i.return}for(;u!==null;){if(i=Et(u),i===null)return;if(s=i.tag,s===5||s===6){r=o=i;continue e}u=u.parentNode}}r=r.return}Ks(function(){var a=o,h=pi(n),m=[];e:{var p=ka.get(e);if(p!==void 0){var w=wi,S=e;switch(e){case"keypress":if(Lr(n)===0)break e;case"keydown":case"keyup":w=ld;break;case"focusin":S="focus",w=Fl;break;case"focusout":S="blur",w=Fl;break;case"beforeblur":case"afterblur":w=Fl;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":w=yu;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":w=Kf;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":w=ud;break;case ya:case ga:case wa:w=Gf;break;case Sa:w=ad;break;case"scroll":w=Hf;break;case"wheel":w=fd;break;case"copy":case"cut":case"paste":w=Jf;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":w=wu}var g=(t&4)!==0,L=!g&&e==="scroll",f=g?p!==null?p+"Capture":null:p;g=[];for(var c=a,d;c!==null;){d=c;var v=d.stateNode;if(d.tag===5&&v!==null&&(d=v,f!==null&&(v=Un(c,f),v!=null&&g.push(Qn(c,v,d)))),L)break;c=c.return}0<g.length&&(p=new w(p,S,null,n,h),m.push({event:p,listeners:g}))}}if(!(t&7)){e:{if(p=e==="mouseover"||e==="pointerover",w=e==="mouseout"||e==="pointerout",p&&n!==So&&(S=n.relatedTarget||n.fromElement)&&(Et(S)||S[Ye]))break e;if((w||p)&&(p=h.window===h?h:(p=h.ownerDocument)?p.defaultView||p.parentWindow:window,w?(S=n.relatedTarget||n.toElement,w=a,S=S?Et(S):null,S!==null&&(L=jt(S),S!==L||S.tag!==5&&S.tag!==6)&&(S=null)):(w=null,S=a),w!==S)){if(g=yu,v="onMouseLeave",f="onMouseEnter",c="mouse",(e==="pointerout"||e==="pointerover")&&(g=wu,v="onPointerLeave",f="onPointerEnter",c="pointer"),L=w==null?p:Vt(w),d=S==null?p:Vt(S),p=new g(v,c+"leave",w,n,h),p.target=L,p.relatedTarget=d,v=null,Et(h)===a&&(g=new g(f,c+"enter",S,n,h),g.target=d,g.relatedTarget=L,v=g),L=v,w&&S)t:{for(g=w,f=S,c=0,d=g;d;d=Mt(d))c++;for(d=0,v=f;v;v=Mt(v))d++;for(;0<c-d;)g=Mt(g),c--;for(;0<d-c;)f=Mt(f),d--;for(;c--;){if(g===f||f!==null&&g===f.alternate)break t;g=Mt(g),f=Mt(f)}g=null}else g=null;w!==null&&Ru(m,p,w,g,!1),S!==null&&L!==null&&Ru(m,L,S,g,!0)}}e:{if(p=a?Vt(a):window,w=p.nodeName&&p.nodeName.toLowerCase(),w==="select"||w==="input"&&p.type==="file")var E=gd;else if(Eu(p))if(da)E=Ed;else{E=Sd;var _=wd}else(w=p.nodeName)&&w.toLowerCase()==="input"&&(p.type==="checkbox"||p.type==="radio")&&(E=kd);if(E&&(E=E(e,a))){fa(m,E,n,h);break e}_&&_(e,p,a),e==="focusout"&&(_=p._wrapperState)&&_.controlled&&p.type==="number"&&mo(p,"number",p.value)}switch(_=a?Vt(a):window,e){case"focusin":(Eu(_)||_.contentEditable==="true")&&($t=_,No=a,Tn=null);break;case"focusout":Tn=No=$t=null;break;case"mousedown":Lo=!0;break;case"contextmenu":case"mouseup":case"dragend":Lo=!1,Nu(m,n,h);break;case"selectionchange":if(_d)break;case"keydown":case"keyup":Nu(m,n,h)}var P;if(ki)e:{switch(e){case"compositionstart":var N="onCompositionStart";break e;case"compositionend":N="onCompositionEnd";break e;case"compositionupdate":N="onCompositionUpdate";break e}N=void 0}else Ut?aa(e,n)&&(N="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(N="onCompositionStart");N&&(sa&&n.locale!=="ko"&&(Ut||N!=="onCompositionStart"?N==="onCompositionEnd"&&Ut&&(P=ua()):(nt=h,gi="value"in nt?nt.value:nt.textContent,Ut=!0)),_=Hr(a,N),0<_.length&&(N=new gu(N,e,null,n,h),m.push({event:N,listeners:_}),P?N.data=P:(P=ca(n),P!==null&&(N.data=P)))),(P=pd?hd(e,n):md(e,n))&&(a=Hr(a,"onBeforeInput"),0<a.length&&(h=new gu("onBeforeInput","beforeinput",null,n,h),m.push({event:h,listeners:a}),h.data=P))}Ea(m,t)})}function Qn(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Hr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,o=l.stateNode;l.tag===5&&o!==null&&(l=o,o=Un(e,n),o!=null&&r.unshift(Qn(e,o,l)),o=Un(e,t),o!=null&&r.push(Qn(e,o,l))),e=e.return}return r}function Mt(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Ru(e,t,n,r,l){for(var o=t._reactName,i=[];n!==null&&n!==r;){var u=n,s=u.alternate,a=u.stateNode;if(s!==null&&s===r)break;u.tag===5&&a!==null&&(u=a,l?(s=Un(n,o),s!=null&&i.unshift(Qn(n,s,u))):l||(s=Un(n,o),s!=null&&i.push(Qn(n,s,u)))),n=n.return}i.length!==0&&e.push({event:t,listeners:i})}var zd=/\r\n?/g,Rd=/\u0000|\uFFFD/g;function Tu(e){return(typeof e=="string"?e:""+e).replace(zd,`
+`).replace(Rd,"")}function yr(e,t,n){if(t=Tu(t),Tu(e)!==t&&n)throw Error(y(425))}function Qr(){}var zo=null,Ro=null;function To(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Oo=typeof setTimeout=="function"?setTimeout:void 0,Td=typeof clearTimeout=="function"?clearTimeout:void 0,Ou=typeof Promise=="function"?Promise:void 0,Od=typeof queueMicrotask=="function"?queueMicrotask:typeof Ou<"u"?function(e){return Ou.resolve(null).then(e).catch(Id)}:Oo;function Id(e){setTimeout(function(){throw e})}function Ql(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),Vn(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);Vn(t)}function st(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function Iu(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var cn=Math.random().toString(36).slice(2),Ue="__reactFiber$"+cn,Kn="__reactProps$"+cn,Ye="__reactContainer$"+cn,Io="__reactEvents$"+cn,jd="__reactListeners$"+cn,Md="__reactHandles$"+cn;function Et(e){var t=e[Ue];if(t)return t;for(var n=e.parentNode;n;){if(t=n[Ye]||n[Ue]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=Iu(e);e!==null;){if(n=e[Ue])return n;e=Iu(e)}return t}e=n,n=e.parentNode}return null}function rr(e){return e=e[Ue]||e[Ye],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function Vt(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(y(33))}function pl(e){return e[Kn]||null}var jo=[],Wt=-1;function vt(e){return{current:e}}function F(e){0>Wt||(e.current=jo[Wt],jo[Wt]=null,Wt--)}function M(e,t){Wt++,jo[Wt]=e.current,e.current=t}var ht={},ie=vt(ht),pe=vt(!1),Lt=ht;function tn(e,t){var n=e.type.contextTypes;if(!n)return ht;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},o;for(o in n)l[o]=t[o];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function he(e){return e=e.childContextTypes,e!=null}function Kr(){F(pe),F(ie)}function ju(e,t,n){if(ie.current!==ht)throw Error(y(168));M(ie,t),M(pe,n)}function Ca(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(y(108,wf(e)||"Unknown",l));return V({},n,r)}function Yr(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||ht,Lt=ie.current,M(ie,e),M(pe,pe.current),!0}function Mu(e,t,n){var r=e.stateNode;if(!r)throw Error(y(169));n?(e=Ca(e,t,Lt),r.__reactInternalMemoizedMergedChildContext=e,F(pe),F(ie),M(ie,e)):F(pe),M(pe,n)}var We=null,hl=!1,Kl=!1;function _a(e){We===null?We=[e]:We.push(e)}function Dd(e){hl=!0,_a(e)}function yt(){if(!Kl&&We!==null){Kl=!0;var e=0,t=j;try{var n=We;for(j=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}We=null,hl=!1}catch(l){throw We!==null&&(We=We.slice(e+1)),Zs(hi,yt),l}finally{j=t,Kl=!1}}return null}var At=[],Ht=0,Xr=null,Gr=0,Ee=[],xe=0,zt=null,Ae=1,He="";function St(e,t){At[Ht++]=Gr,At[Ht++]=Xr,Xr=e,Gr=t}function Pa(e,t,n){Ee[xe++]=Ae,Ee[xe++]=He,Ee[xe++]=zt,zt=e;var r=Ae;e=He;var l=32-Ie(r)-1;r&=~(1<<l),n+=1;var o=32-Ie(t)+l;if(30<o){var i=l-l%5;o=(r&(1<<i)-1).toString(32),r>>=i,l-=i,Ae=1<<32-Ie(t)+l|n<<l|r,He=o+e}else Ae=1<<o|n<<l|r,He=e}function xi(e){e.return!==null&&(St(e,1),Pa(e,1,0))}function Ci(e){for(;e===Xr;)Xr=At[--Ht],At[Ht]=null,Gr=At[--Ht],At[Ht]=null;for(;e===zt;)zt=Ee[--xe],Ee[xe]=null,He=Ee[--xe],Ee[xe]=null,Ae=Ee[--xe],Ee[xe]=null}var ge=null,ye=null,U=!1,Oe=null;function Na(e,t){var n=Ce(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function Du(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,ge=e,ye=st(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,ge=e,ye=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=zt!==null?{id:Ae,overflow:He}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Ce(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,ge=e,ye=null,!0):!1;default:return!1}}function Mo(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Do(e){if(U){var t=ye;if(t){var n=t;if(!Du(e,t)){if(Mo(e))throw Error(y(418));t=st(n.nextSibling);var r=ge;t&&Du(e,t)?Na(r,n):(e.flags=e.flags&-4097|2,U=!1,ge=e)}}else{if(Mo(e))throw Error(y(418));e.flags=e.flags&-4097|2,U=!1,ge=e}}}function Fu(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;ge=e}function gr(e){if(e!==ge)return!1;if(!U)return Fu(e),U=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!To(e.type,e.memoizedProps)),t&&(t=ye)){if(Mo(e))throw La(),Error(y(418));for(;t;)Na(e,t),t=st(t.nextSibling)}if(Fu(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(y(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){ye=st(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}ye=null}}else ye=ge?st(e.stateNode.nextSibling):null;return!0}function La(){for(var e=ye;e;)e=st(e.nextSibling)}function nn(){ye=ge=null,U=!1}function _i(e){Oe===null?Oe=[e]:Oe.push(e)}var Fd=Ze.ReactCurrentBatchConfig;function Re(e,t){if(e&&e.defaultProps){t=V({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}var Zr=vt(null),Jr=null,Qt=null,Pi=null;function Ni(){Pi=Qt=Jr=null}function Li(e){var t=Zr.current;F(Zr),e._currentValue=t}function Fo(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function qt(e,t){Jr=e,Pi=Qt=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(de=!0),e.firstContext=null)}function Pe(e){var t=e._currentValue;if(Pi!==e)if(e={context:e,memoizedValue:t,next:null},Qt===null){if(Jr===null)throw Error(y(308));Qt=e,Jr.dependencies={lanes:0,firstContext:e}}else Qt=Qt.next=e;return t}var xt=null;function zi(e){xt===null?xt=[e]:xt.push(e)}function za(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,zi(t)):(n.next=l.next,l.next=n),t.interleaved=n,Xe(e,r)}function Xe(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var be=!1;function Ri(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Ra(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function Qe(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function at(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,I&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,Xe(e,n)}return l=r.interleaved,l===null?(t.next=t,zi(r)):(t.next=l.next,l.next=t),r.interleaved=t,Xe(e,n)}function zr(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mi(e,n)}}function Uu(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,o=null;if(n=n.firstBaseUpdate,n!==null){do{var i={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};o===null?l=o=i:o=o.next=i,n=n.next}while(n!==null);o===null?l=o=t:o=o.next=t}else l=o=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:o,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function qr(e,t,n,r){var l=e.updateQueue;be=!1;var o=l.firstBaseUpdate,i=l.lastBaseUpdate,u=l.shared.pending;if(u!==null){l.shared.pending=null;var s=u,a=s.next;s.next=null,i===null?o=a:i.next=a,i=s;var h=e.alternate;h!==null&&(h=h.updateQueue,u=h.lastBaseUpdate,u!==i&&(u===null?h.firstBaseUpdate=a:u.next=a,h.lastBaseUpdate=s))}if(o!==null){var m=l.baseState;i=0,h=a=s=null,u=o;do{var p=u.lane,w=u.eventTime;if((r&p)===p){h!==null&&(h=h.next={eventTime:w,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var S=e,g=u;switch(p=t,w=n,g.tag){case 1:if(S=g.payload,typeof S=="function"){m=S.call(w,m,p);break e}m=S;break e;case 3:S.flags=S.flags&-65537|128;case 0:if(S=g.payload,p=typeof S=="function"?S.call(w,m,p):S,p==null)break e;m=V({},m,p);break e;case 2:be=!0}}u.callback!==null&&u.lane!==0&&(e.flags|=64,p=l.effects,p===null?l.effects=[u]:p.push(u))}else w={eventTime:w,lane:p,tag:u.tag,payload:u.payload,callback:u.callback,next:null},h===null?(a=h=w,s=m):h=h.next=w,i|=p;if(u=u.next,u===null){if(u=l.shared.pending,u===null)break;p=u,u=p.next,p.next=null,l.lastBaseUpdate=p,l.shared.pending=null}}while(1);if(h===null&&(s=m),l.baseState=s,l.firstBaseUpdate=a,l.lastBaseUpdate=h,t=l.shared.interleaved,t!==null){l=t;do i|=l.lane,l=l.next;while(l!==t)}else o===null&&(l.shared.lanes=0);Tt|=i,e.lanes=i,e.memoizedState=m}}function $u(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(y(191,l));l.call(r)}}}var Ta=new zs.Component().refs;function Uo(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:V({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var ml={isMounted:function(e){return(e=e._reactInternals)?jt(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=se(),l=ft(e),o=Qe(r,l);o.payload=t,n!=null&&(o.callback=n),t=at(e,o,l),t!==null&&(je(t,e,l,r),zr(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=se(),l=ft(e),o=Qe(r,l);o.tag=1,o.payload=t,n!=null&&(o.callback=n),t=at(e,o,l),t!==null&&(je(t,e,l,r),zr(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=se(),r=ft(e),l=Qe(n,r);l.tag=2,t!=null&&(l.callback=t),t=at(e,l,r),t!==null&&(je(t,e,r,n),zr(t,e,r))}};function Bu(e,t,n,r,l,o,i){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,o,i):t.prototype&&t.prototype.isPureReactComponent?!An(n,r)||!An(l,o):!0}function Oa(e,t,n){var r=!1,l=ht,o=t.contextType;return typeof o=="object"&&o!==null?o=Pe(o):(l=he(t)?Lt:ie.current,r=t.contextTypes,o=(r=r!=null)?tn(e,l):ht),t=new t(n,o),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=ml,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=o),t}function Vu(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&ml.enqueueReplaceState(t,t.state,null)}function $o(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs=Ta,Ri(e);var o=t.contextType;typeof o=="object"&&o!==null?l.context=Pe(o):(o=he(t)?Lt:ie.current,l.context=tn(e,o)),l.state=e.memoizedState,o=t.getDerivedStateFromProps,typeof o=="function"&&(Uo(e,t,o,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&ml.enqueueReplaceState(l,l.state,null),qr(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function Sn(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(y(309));var r=n.stateNode}if(!r)throw Error(y(147,e));var l=r,o=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===o?t.ref:(t=function(i){var u=l.refs;u===Ta&&(u=l.refs={}),i===null?delete u[o]:u[o]=i},t._stringRef=o,t)}if(typeof e!="string")throw Error(y(284));if(!n._owner)throw Error(y(290,e))}return e}function wr(e,t){throw e=Object.prototype.toString.call(t),Error(y(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Wu(e){var t=e._init;return t(e._payload)}function Ia(e){function t(f,c){if(e){var d=f.deletions;d===null?(f.deletions=[c],f.flags|=16):d.push(c)}}function n(f,c){if(!e)return null;for(;c!==null;)t(f,c),c=c.sibling;return null}function r(f,c){for(f=new Map;c!==null;)c.key!==null?f.set(c.key,c):f.set(c.index,c),c=c.sibling;return f}function l(f,c){return f=dt(f,c),f.index=0,f.sibling=null,f}function o(f,c,d){return f.index=d,e?(d=f.alternate,d!==null?(d=d.index,d<c?(f.flags|=2,c):d):(f.flags|=2,c)):(f.flags|=1048576,c)}function i(f){return e&&f.alternate===null&&(f.flags|=2),f}function u(f,c,d,v){return c===null||c.tag!==6?(c=bl(d,f.mode,v),c.return=f,c):(c=l(c,d),c.return=f,c)}function s(f,c,d,v){var E=d.type;return E===Ft?h(f,c,d.props.children,v,d.key):c!==null&&(c.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===qe&&Wu(E)===c.type)?(v=l(c,d.props),v.ref=Sn(f,c,d),v.return=f,v):(v=Mr(d.type,d.key,d.props,null,f.mode,v),v.ref=Sn(f,c,d),v.return=f,v)}function a(f,c,d,v){return c===null||c.tag!==4||c.stateNode.containerInfo!==d.containerInfo||c.stateNode.implementation!==d.implementation?(c=eo(d,f.mode,v),c.return=f,c):(c=l(c,d.children||[]),c.return=f,c)}function h(f,c,d,v,E){return c===null||c.tag!==7?(c=Pt(d,f.mode,v,E),c.return=f,c):(c=l(c,d),c.return=f,c)}function m(f,c,d){if(typeof c=="string"&&c!==""||typeof c=="number")return c=bl(""+c,f.mode,d),c.return=f,c;if(typeof c=="object"&&c!==null){switch(c.$$typeof){case sr:return d=Mr(c.type,c.key,c.props,null,f.mode,d),d.ref=Sn(f,null,c),d.return=f,d;case Dt:return c=eo(c,f.mode,d),c.return=f,c;case qe:var v=c._init;return m(f,v(c._payload),d)}if(Cn(c)||mn(c))return c=Pt(c,f.mode,d,null),c.return=f,c;wr(f,c)}return null}function p(f,c,d,v){var E=c!==null?c.key:null;if(typeof d=="string"&&d!==""||typeof d=="number")return E!==null?null:u(f,c,""+d,v);if(typeof d=="object"&&d!==null){switch(d.$$typeof){case sr:return d.key===E?s(f,c,d,v):null;case Dt:return d.key===E?a(f,c,d,v):null;case qe:return E=d._init,p(f,c,E(d._payload),v)}if(Cn(d)||mn(d))return E!==null?null:h(f,c,d,v,null);wr(f,d)}return null}function w(f,c,d,v,E){if(typeof v=="string"&&v!==""||typeof v=="number")return f=f.get(d)||null,u(c,f,""+v,E);if(typeof v=="object"&&v!==null){switch(v.$$typeof){case sr:return f=f.get(v.key===null?d:v.key)||null,s(c,f,v,E);case Dt:return f=f.get(v.key===null?d:v.key)||null,a(c,f,v,E);case qe:var _=v._init;return w(f,c,d,_(v._payload),E)}if(Cn(v)||mn(v))return f=f.get(d)||null,h(c,f,v,E,null);wr(c,v)}return null}function S(f,c,d,v){for(var E=null,_=null,P=c,N=c=0,A=null;P!==null&&N<d.length;N++){P.index>N?(A=P,P=null):A=P.sibling;var O=p(f,P,d[N],v);if(O===null){P===null&&(P=A);break}e&&P&&O.alternate===null&&t(f,P),c=o(O,c,N),_===null?E=O:_.sibling=O,_=O,P=A}if(N===d.length)return n(f,P),U&&St(f,N),E;if(P===null){for(;N<d.length;N++)P=m(f,d[N],v),P!==null&&(c=o(P,c,N),_===null?E=P:_.sibling=P,_=P);return U&&St(f,N),E}for(P=r(f,P);N<d.length;N++)A=w(P,f,N,d[N],v),A!==null&&(e&&A.alternate!==null&&P.delete(A.key===null?N:A.key),c=o(A,c,N),_===null?E=A:_.sibling=A,_=A);return e&&P.forEach(function(Le){return t(f,Le)}),U&&St(f,N),E}function g(f,c,d,v){var E=mn(d);if(typeof E!="function")throw Error(y(150));if(d=E.call(d),d==null)throw Error(y(151));for(var _=E=null,P=c,N=c=0,A=null,O=d.next();P!==null&&!O.done;N++,O=d.next()){P.index>N?(A=P,P=null):A=P.sibling;var Le=p(f,P,O.value,v);if(Le===null){P===null&&(P=A);break}e&&P&&Le.alternate===null&&t(f,P),c=o(Le,c,N),_===null?E=Le:_.sibling=Le,_=Le,P=A}if(O.done)return n(f,P),U&&St(f,N),E;if(P===null){for(;!O.done;N++,O=d.next())O=m(f,O.value,v),O!==null&&(c=o(O,c,N),_===null?E=O:_.sibling=O,_=O);return U&&St(f,N),E}for(P=r(f,P);!O.done;N++,O=d.next())O=w(P,f,N,O.value,v),O!==null&&(e&&O.alternate!==null&&P.delete(O.key===null?N:O.key),c=o(O,c,N),_===null?E=O:_.sibling=O,_=O);return e&&P.forEach(function(pn){return t(f,pn)}),U&&St(f,N),E}function L(f,c,d,v){if(typeof d=="object"&&d!==null&&d.type===Ft&&d.key===null&&(d=d.props.children),typeof d=="object"&&d!==null){switch(d.$$typeof){case sr:e:{for(var E=d.key,_=c;_!==null;){if(_.key===E){if(E=d.type,E===Ft){if(_.tag===7){n(f,_.sibling),c=l(_,d.props.children),c.return=f,f=c;break e}}else if(_.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===qe&&Wu(E)===_.type){n(f,_.sibling),c=l(_,d.props),c.ref=Sn(f,_,d),c.return=f,f=c;break e}n(f,_);break}else t(f,_);_=_.sibling}d.type===Ft?(c=Pt(d.props.children,f.mode,v,d.key),c.return=f,f=c):(v=Mr(d.type,d.key,d.props,null,f.mode,v),v.ref=Sn(f,c,d),v.return=f,f=v)}return i(f);case Dt:e:{for(_=d.key;c!==null;){if(c.key===_)if(c.tag===4&&c.stateNode.containerInfo===d.containerInfo&&c.stateNode.implementation===d.implementation){n(f,c.sibling),c=l(c,d.children||[]),c.return=f,f=c;break e}else{n(f,c);break}else t(f,c);c=c.sibling}c=eo(d,f.mode,v),c.return=f,f=c}return i(f);case qe:return _=d._init,L(f,c,_(d._payload),v)}if(Cn(d))return S(f,c,d,v);if(mn(d))return g(f,c,d,v);wr(f,d)}return typeof d=="string"&&d!==""||typeof d=="number"?(d=""+d,c!==null&&c.tag===6?(n(f,c.sibling),c=l(c,d),c.return=f,f=c):(n(f,c),c=bl(d,f.mode,v),c.return=f,f=c),i(f)):n(f,c)}return L}var rn=Ia(!0),ja=Ia(!1),lr={},Be=vt(lr),Yn=vt(lr),Xn=vt(lr);function Ct(e){if(e===lr)throw Error(y(174));return e}function Ti(e,t){switch(M(Xn,t),M(Yn,e),M(Be,lr),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:yo(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=yo(t,e)}F(Be),M(Be,t)}function ln(){F(Be),F(Yn),F(Xn)}function Ma(e){Ct(Xn.current);var t=Ct(Be.current),n=yo(t,e.type);t!==n&&(M(Yn,e),M(Be,n))}function Oi(e){Yn.current===e&&(F(Be),F(Yn))}var $=vt(0);function br(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var Yl=[];function Ii(){for(var e=0;e<Yl.length;e++)Yl[e]._workInProgressVersionPrimary=null;Yl.length=0}var Rr=Ze.ReactCurrentDispatcher,Xl=Ze.ReactCurrentBatchConfig,Rt=0,B=null,X=null,J=null,el=!1,On=!1,Gn=0,Ud=0;function re(){throw Error(y(321))}function ji(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Me(e[n],t[n]))return!1;return!0}function Mi(e,t,n,r,l,o){if(Rt=o,B=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,Rr.current=e===null||e.memoizedState===null?Wd:Ad,e=n(r,l),On){o=0;do{if(On=!1,Gn=0,25<=o)throw Error(y(301));o+=1,J=X=null,t.updateQueue=null,Rr.current=Hd,e=n(r,l)}while(On)}if(Rr.current=tl,t=X!==null&&X.next!==null,Rt=0,J=X=B=null,el=!1,t)throw Error(y(300));return e}function Di(){var e=Gn!==0;return Gn=0,e}function Fe(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return J===null?B.memoizedState=J=e:J=J.next=e,J}function Ne(){if(X===null){var e=B.alternate;e=e!==null?e.memoizedState:null}else e=X.next;var t=J===null?B.memoizedState:J.next;if(t!==null)J=t,X=e;else{if(e===null)throw Error(y(310));X=e,e={memoizedState:X.memoizedState,baseState:X.baseState,baseQueue:X.baseQueue,queue:X.queue,next:null},J===null?B.memoizedState=J=e:J=J.next=e}return J}function Zn(e,t){return typeof t=="function"?t(e):t}function Gl(e){var t=Ne(),n=t.queue;if(n===null)throw Error(y(311));n.lastRenderedReducer=e;var r=X,l=r.baseQueue,o=n.pending;if(o!==null){if(l!==null){var i=l.next;l.next=o.next,o.next=i}r.baseQueue=l=o,n.pending=null}if(l!==null){o=l.next,r=r.baseState;var u=i=null,s=null,a=o;do{var h=a.lane;if((Rt&h)===h)s!==null&&(s=s.next={lane:0,action:a.action,hasEagerState:a.hasEagerState,eagerState:a.eagerState,next:null}),r=a.hasEagerState?a.eagerState:e(r,a.action);else{var m={lane:h,action:a.action,hasEagerState:a.hasEagerState,eagerState:a.eagerState,next:null};s===null?(u=s=m,i=r):s=s.next=m,B.lanes|=h,Tt|=h}a=a.next}while(a!==null&&a!==o);s===null?i=r:s.next=u,Me(r,t.memoizedState)||(de=!0),t.memoizedState=r,t.baseState=i,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do o=l.lane,B.lanes|=o,Tt|=o,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function Zl(e){var t=Ne(),n=t.queue;if(n===null)throw Error(y(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,o=t.memoizedState;if(l!==null){n.pending=null;var i=l=l.next;do o=e(o,i.action),i=i.next;while(i!==l);Me(o,t.memoizedState)||(de=!0),t.memoizedState=o,t.baseQueue===null&&(t.baseState=o),n.lastRenderedState=o}return[o,r]}function Da(){}function Fa(e,t){var n=B,r=Ne(),l=t(),o=!Me(r.memoizedState,l);if(o&&(r.memoizedState=l,de=!0),r=r.queue,Fi(Ba.bind(null,n,r,e),[e]),r.getSnapshot!==t||o||J!==null&&J.memoizedState.tag&1){if(n.flags|=2048,Jn(9,$a.bind(null,n,r,l,t),void 0,null),q===null)throw Error(y(349));Rt&30||Ua(n,t,l)}return l}function Ua(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=B.updateQueue,t===null?(t={lastEffect:null,stores:null},B.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function $a(e,t,n,r){t.value=n,t.getSnapshot=r,Va(t)&&Wa(e)}function Ba(e,t,n){return n(function(){Va(t)&&Wa(e)})}function Va(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Me(e,n)}catch{return!0}}function Wa(e){var t=Xe(e,1);t!==null&&je(t,e,1,-1)}function Au(e){var t=Fe();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:Zn,lastRenderedState:e},t.queue=e,e=e.dispatch=Vd.bind(null,B,e),[t.memoizedState,e]}function Jn(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=B.updateQueue,t===null?(t={lastEffect:null,stores:null},B.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function Aa(){return Ne().memoizedState}function Tr(e,t,n,r){var l=Fe();B.flags|=e,l.memoizedState=Jn(1|t,n,void 0,r===void 0?null:r)}function vl(e,t,n,r){var l=Ne();r=r===void 0?null:r;var o=void 0;if(X!==null){var i=X.memoizedState;if(o=i.destroy,r!==null&&ji(r,i.deps)){l.memoizedState=Jn(t,n,o,r);return}}B.flags|=e,l.memoizedState=Jn(1|t,n,o,r)}function Hu(e,t){return Tr(8390656,8,e,t)}function Fi(e,t){return vl(2048,8,e,t)}function Ha(e,t){return vl(4,2,e,t)}function Qa(e,t){return vl(4,4,e,t)}function Ka(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function Ya(e,t,n){return n=n!=null?n.concat([e]):null,vl(4,4,Ka.bind(null,t,e),n)}function Ui(){}function Xa(e,t){var n=Ne();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ji(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Ga(e,t){var n=Ne();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&ji(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Za(e,t,n){return Rt&21?(Me(n,t)||(n=bs(),B.lanes|=n,Tt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,de=!0),e.memoizedState=n)}function $d(e,t){var n=j;j=n!==0&&4>n?n:4,e(!0);var r=Xl.transition;Xl.transition={};try{e(!1),t()}finally{j=n,Xl.transition=r}}function Ja(){return Ne().memoizedState}function Bd(e,t,n){var r=ft(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},qa(e))ba(t,n);else if(n=za(e,t,n,r),n!==null){var l=se();je(n,e,r,l),ec(n,t,r)}}function Vd(e,t,n){var r=ft(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(qa(e))ba(t,l);else{var o=e.alternate;if(e.lanes===0&&(o===null||o.lanes===0)&&(o=t.lastRenderedReducer,o!==null))try{var i=t.lastRenderedState,u=o(i,n);if(l.hasEagerState=!0,l.eagerState=u,Me(u,i)){var s=t.interleaved;s===null?(l.next=l,zi(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=za(e,t,l,r),n!==null&&(l=se(),je(n,e,r,l),ec(n,t,r))}}function qa(e){var t=e.alternate;return e===B||t!==null&&t===B}function ba(e,t){On=el=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function ec(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,mi(e,n)}}var tl={readContext:Pe,useCallback:re,useContext:re,useEffect:re,useImperativeHandle:re,useInsertionEffect:re,useLayoutEffect:re,useMemo:re,useReducer:re,useRef:re,useState:re,useDebugValue:re,useDeferredValue:re,useTransition:re,useMutableSource:re,useSyncExternalStore:re,useId:re,unstable_isNewReconciler:!1},Wd={readContext:Pe,useCallback:function(e,t){return Fe().memoizedState=[e,t===void 0?null:t],e},useContext:Pe,useEffect:Hu,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,Tr(4194308,4,Ka.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Tr(4194308,4,e,t)},useInsertionEffect:function(e,t){return Tr(4,2,e,t)},useMemo:function(e,t){var n=Fe();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=Fe();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Bd.bind(null,B,e),[r.memoizedState,e]},useRef:function(e){var t=Fe();return e={current:e},t.memoizedState=e},useState:Au,useDebugValue:Ui,useDeferredValue:function(e){return Fe().memoizedState=e},useTransition:function(){var e=Au(!1),t=e[0];return e=$d.bind(null,e[1]),Fe().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=B,l=Fe();if(U){if(n===void 0)throw Error(y(407));n=n()}else{if(n=t(),q===null)throw Error(y(349));Rt&30||Ua(r,t,n)}l.memoizedState=n;var o={value:n,getSnapshot:t};return l.queue=o,Hu(Ba.bind(null,r,o,e),[e]),r.flags|=2048,Jn(9,$a.bind(null,r,o,n,t),void 0,null),n},useId:function(){var e=Fe(),t=q.identifierPrefix;if(U){var n=He,r=Ae;n=(r&~(1<<32-Ie(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=Gn++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Ud++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Ad={readContext:Pe,useCallback:Xa,useContext:Pe,useEffect:Fi,useImperativeHandle:Ya,useInsertionEffect:Ha,useLayoutEffect:Qa,useMemo:Ga,useReducer:Gl,useRef:Aa,useState:function(){return Gl(Zn)},useDebugValue:Ui,useDeferredValue:function(e){var t=Ne();return Za(t,X.memoizedState,e)},useTransition:function(){var e=Gl(Zn)[0],t=Ne().memoizedState;return[e,t]},useMutableSource:Da,useSyncExternalStore:Fa,useId:Ja,unstable_isNewReconciler:!1},Hd={readContext:Pe,useCallback:Xa,useContext:Pe,useEffect:Fi,useImperativeHandle:Ya,useInsertionEffect:Ha,useLayoutEffect:Qa,useMemo:Ga,useReducer:Zl,useRef:Aa,useState:function(){return Zl(Zn)},useDebugValue:Ui,useDeferredValue:function(e){var t=Ne();return X===null?t.memoizedState=e:Za(t,X.memoizedState,e)},useTransition:function(){var e=Zl(Zn)[0],t=Ne().memoizedState;return[e,t]},useMutableSource:Da,useSyncExternalStore:Fa,useId:Ja,unstable_isNewReconciler:!1};function on(e,t){try{var n="",r=t;do n+=gf(r),r=r.return;while(r);var l=n}catch(o){l=`
+Error generating stack: `+o.message+`
+`+o.stack}return{value:e,source:t,stack:l,digest:null}}function Jl(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function Bo(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Qd=typeof WeakMap=="function"?WeakMap:Map;function tc(e,t,n){n=Qe(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){rl||(rl=!0,Zo=r),Bo(e,t)},n}function nc(e,t,n){n=Qe(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){Bo(e,t)}}var o=e.stateNode;return o!==null&&typeof o.componentDidCatch=="function"&&(n.callback=function(){Bo(e,t),typeof r!="function"&&(ct===null?ct=new Set([this]):ct.add(this));var i=t.stack;this.componentDidCatch(t.value,{componentStack:i!==null?i:""})}),n}function Qu(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Qd;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=op.bind(null,e,t,n),t.then(e,e))}function Ku(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function Yu(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=Qe(-1,1),t.tag=2,at(n,t,1))),n.lanes|=1),e)}var Kd=Ze.ReactCurrentOwner,de=!1;function ue(e,t,n,r){t.child=e===null?ja(t,null,n,r):rn(t,e.child,n,r)}function Xu(e,t,n,r,l){n=n.render;var o=t.ref;return qt(t,l),r=Mi(e,t,n,r,o,l),n=Di(),e!==null&&!de?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Ge(e,t,l)):(U&&n&&xi(t),t.flags|=1,ue(e,t,r,l),t.child)}function Gu(e,t,n,r,l){if(e===null){var o=n.type;return typeof o=="function"&&!Ki(o)&&o.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=o,rc(e,t,o,r,l)):(e=Mr(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(o=e.child,!(e.lanes&l)){var i=o.memoizedProps;if(n=n.compare,n=n!==null?n:An,n(i,r)&&e.ref===t.ref)return Ge(e,t,l)}return t.flags|=1,e=dt(o,r),e.ref=t.ref,e.return=t,t.child=e}function rc(e,t,n,r,l){if(e!==null){var o=e.memoizedProps;if(An(o,r)&&e.ref===t.ref)if(de=!1,t.pendingProps=r=o,(e.lanes&l)!==0)e.flags&131072&&(de=!0);else return t.lanes=e.lanes,Ge(e,t,l)}return Vo(e,t,n,r,l)}function lc(e,t,n){var r=t.pendingProps,l=r.children,o=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},M(Yt,ve),ve|=n;else{if(!(n&1073741824))return e=o!==null?o.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,M(Yt,ve),ve|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=o!==null?o.baseLanes:n,M(Yt,ve),ve|=r}else o!==null?(r=o.baseLanes|n,t.memoizedState=null):r=n,M(Yt,ve),ve|=r;return ue(e,t,l,n),t.child}function oc(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function Vo(e,t,n,r,l){var o=he(n)?Lt:ie.current;return o=tn(t,o),qt(t,l),n=Mi(e,t,n,r,o,l),r=Di(),e!==null&&!de?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Ge(e,t,l)):(U&&r&&xi(t),t.flags|=1,ue(e,t,n,l),t.child)}function Zu(e,t,n,r,l){if(he(n)){var o=!0;Yr(t)}else o=!1;if(qt(t,l),t.stateNode===null)Or(e,t),Oa(t,n,r),$o(t,n,r,l),r=!0;else if(e===null){var i=t.stateNode,u=t.memoizedProps;i.props=u;var s=i.context,a=n.contextType;typeof a=="object"&&a!==null?a=Pe(a):(a=he(n)?Lt:ie.current,a=tn(t,a));var h=n.getDerivedStateFromProps,m=typeof h=="function"||typeof i.getSnapshotBeforeUpdate=="function";m||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==r||s!==a)&&Vu(t,i,r,a),be=!1;var p=t.memoizedState;i.state=p,qr(t,r,i,l),s=t.memoizedState,u!==r||p!==s||pe.current||be?(typeof h=="function"&&(Uo(t,n,h,r),s=t.memoizedState),(u=be||Bu(t,n,u,r,p,s,a))?(m||typeof i.UNSAFE_componentWillMount!="function"&&typeof i.componentWillMount!="function"||(typeof i.componentWillMount=="function"&&i.componentWillMount(),typeof i.UNSAFE_componentWillMount=="function"&&i.UNSAFE_componentWillMount()),typeof i.componentDidMount=="function"&&(t.flags|=4194308)):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),i.props=r,i.state=s,i.context=a,r=u):(typeof i.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{i=t.stateNode,Ra(e,t),u=t.memoizedProps,a=t.type===t.elementType?u:Re(t.type,u),i.props=a,m=t.pendingProps,p=i.context,s=n.contextType,typeof s=="object"&&s!==null?s=Pe(s):(s=he(n)?Lt:ie.current,s=tn(t,s));var w=n.getDerivedStateFromProps;(h=typeof w=="function"||typeof i.getSnapshotBeforeUpdate=="function")||typeof i.UNSAFE_componentWillReceiveProps!="function"&&typeof i.componentWillReceiveProps!="function"||(u!==m||p!==s)&&Vu(t,i,r,s),be=!1,p=t.memoizedState,i.state=p,qr(t,r,i,l);var S=t.memoizedState;u!==m||p!==S||pe.current||be?(typeof w=="function"&&(Uo(t,n,w,r),S=t.memoizedState),(a=be||Bu(t,n,a,r,p,S,s)||!1)?(h||typeof i.UNSAFE_componentWillUpdate!="function"&&typeof i.componentWillUpdate!="function"||(typeof i.componentWillUpdate=="function"&&i.componentWillUpdate(r,S,s),typeof i.UNSAFE_componentWillUpdate=="function"&&i.UNSAFE_componentWillUpdate(r,S,s)),typeof i.componentDidUpdate=="function"&&(t.flags|=4),typeof i.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=S),i.props=r,i.state=S,i.context=s,r=a):(typeof i.componentDidUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=4),typeof i.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&p===e.memoizedState||(t.flags|=1024),r=!1)}return Wo(e,t,n,r,o,l)}function Wo(e,t,n,r,l,o){oc(e,t);var i=(t.flags&128)!==0;if(!r&&!i)return l&&Mu(t,n,!1),Ge(e,t,o);r=t.stateNode,Kd.current=t;var u=i&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&i?(t.child=rn(t,e.child,null,o),t.child=rn(t,null,u,o)):ue(e,t,u,o),t.memoizedState=r.state,l&&Mu(t,n,!0),t.child}function ic(e){var t=e.stateNode;t.pendingContext?ju(e,t.pendingContext,t.pendingContext!==t.context):t.context&&ju(e,t.context,!1),Ti(e,t.containerInfo)}function Ju(e,t,n,r,l){return nn(),_i(l),t.flags|=256,ue(e,t,n,r),t.child}var Ao={dehydrated:null,treeContext:null,retryLane:0};function Ho(e){return{baseLanes:e,cachePool:null,transitions:null}}function uc(e,t,n){var r=t.pendingProps,l=$.current,o=!1,i=(t.flags&128)!==0,u;if((u=i)||(u=e!==null&&e.memoizedState===null?!1:(l&2)!==0),u?(o=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),M($,l&1),e===null)return Do(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(i=r.children,e=r.fallback,o?(r=t.mode,o=t.child,i={mode:"hidden",children:i},!(r&1)&&o!==null?(o.childLanes=0,o.pendingProps=i):o=wl(i,r,0,null),e=Pt(e,r,n,null),o.return=t,e.return=t,o.sibling=e,t.child=o,t.child.memoizedState=Ho(n),t.memoizedState=Ao,e):$i(t,i));if(l=e.memoizedState,l!==null&&(u=l.dehydrated,u!==null))return Yd(e,t,i,r,u,l,n);if(o){o=r.fallback,i=t.mode,l=e.child,u=l.sibling;var s={mode:"hidden",children:r.children};return!(i&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=dt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),u!==null?o=dt(u,o):(o=Pt(o,i,n,null),o.flags|=2),o.return=t,r.return=t,r.sibling=o,t.child=r,r=o,o=t.child,i=e.child.memoizedState,i=i===null?Ho(n):{baseLanes:i.baseLanes|n,cachePool:null,transitions:i.transitions},o.memoizedState=i,o.childLanes=e.childLanes&~n,t.memoizedState=Ao,r}return o=e.child,e=o.sibling,r=dt(o,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function $i(e,t){return t=wl({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function Sr(e,t,n,r){return r!==null&&_i(r),rn(t,e.child,null,n),e=$i(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Yd(e,t,n,r,l,o,i){if(n)return t.flags&256?(t.flags&=-257,r=Jl(Error(y(422))),Sr(e,t,i,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(o=r.fallback,l=t.mode,r=wl({mode:"visible",children:r.children},l,0,null),o=Pt(o,l,i,null),o.flags|=2,r.return=t,o.return=t,r.sibling=o,t.child=r,t.mode&1&&rn(t,e.child,null,i),t.child.memoizedState=Ho(i),t.memoizedState=Ao,o);if(!(t.mode&1))return Sr(e,t,i,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var u=r.dgst;return r=u,o=Error(y(419)),r=Jl(o,r,void 0),Sr(e,t,i,r)}if(u=(i&e.childLanes)!==0,de||u){if(r=q,r!==null){switch(i&-i){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|i)?0:l,l!==0&&l!==o.retryLane&&(o.retryLane=l,Xe(e,l),je(r,e,l,-1))}return Qi(),r=Jl(Error(y(421))),Sr(e,t,i,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=ip.bind(null,e),l._reactRetry=t,null):(e=o.treeContext,ye=st(l.nextSibling),ge=t,U=!0,Oe=null,e!==null&&(Ee[xe++]=Ae,Ee[xe++]=He,Ee[xe++]=zt,Ae=e.id,He=e.overflow,zt=t),t=$i(t,r.children),t.flags|=4096,t)}function qu(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),Fo(e.return,t,n)}function ql(e,t,n,r,l){var o=e.memoizedState;o===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(o.isBackwards=t,o.rendering=null,o.renderingStartTime=0,o.last=r,o.tail=n,o.tailMode=l)}function sc(e,t,n){var r=t.pendingProps,l=r.revealOrder,o=r.tail;if(ue(e,t,r.children,n),r=$.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&qu(e,n,t);else if(e.tag===19)qu(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(M($,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&br(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),ql(t,!1,l,n,o);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&br(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}ql(t,!0,n,null,o);break;case"together":ql(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Or(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Ge(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),Tt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(y(153));if(t.child!==null){for(e=t.child,n=dt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=dt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Xd(e,t,n){switch(t.tag){case 3:ic(t),nn();break;case 5:Ma(t);break;case 1:he(t.type)&&Yr(t);break;case 4:Ti(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;M(Zr,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?(M($,$.current&1),t.flags|=128,null):n&t.child.childLanes?uc(e,t,n):(M($,$.current&1),e=Ge(e,t,n),e!==null?e.sibling:null);M($,$.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return sc(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),M($,$.current),r)break;return null;case 22:case 23:return t.lanes=0,lc(e,t,n)}return Ge(e,t,n)}var ac,Qo,cc,fc;ac=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}};Qo=function(){};cc=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,Ct(Be.current);var o=null;switch(n){case"input":l=po(e,l),r=po(e,r),o=[];break;case"select":l=V({},l,{value:void 0}),r=V({},r,{value:void 0}),o=[];break;case"textarea":l=vo(e,l),r=vo(e,r),o=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=Qr)}go(n,r);var i;n=null;for(a in l)if(!r.hasOwnProperty(a)&&l.hasOwnProperty(a)&&l[a]!=null)if(a==="style"){var u=l[a];for(i in u)u.hasOwnProperty(i)&&(n||(n={}),n[i]="")}else a!=="dangerouslySetInnerHTML"&&a!=="children"&&a!=="suppressContentEditableWarning"&&a!=="suppressHydrationWarning"&&a!=="autoFocus"&&(Dn.hasOwnProperty(a)?o||(o=[]):(o=o||[]).push(a,null));for(a in r){var s=r[a];if(u=l!=null?l[a]:void 0,r.hasOwnProperty(a)&&s!==u&&(s!=null||u!=null))if(a==="style")if(u){for(i in u)!u.hasOwnProperty(i)||s&&s.hasOwnProperty(i)||(n||(n={}),n[i]="");for(i in s)s.hasOwnProperty(i)&&u[i]!==s[i]&&(n||(n={}),n[i]=s[i])}else n||(o||(o=[]),o.push(a,n)),n=s;else a==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,u=u?u.__html:void 0,s!=null&&u!==s&&(o=o||[]).push(a,s)):a==="children"?typeof s!="string"&&typeof s!="number"||(o=o||[]).push(a,""+s):a!=="suppressContentEditableWarning"&&a!=="suppressHydrationWarning"&&(Dn.hasOwnProperty(a)?(s!=null&&a==="onScroll"&&D("scroll",e),o||u===s||(o=[])):(o=o||[]).push(a,s))}n&&(o=o||[]).push("style",n);var a=o;(t.updateQueue=a)&&(t.flags|=4)}};fc=function(e,t,n,r){n!==r&&(t.flags|=4)};function kn(e,t){if(!U)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function le(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function Gd(e,t,n){var r=t.pendingProps;switch(Ci(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return le(t),null;case 1:return he(t.type)&&Kr(),le(t),null;case 3:return r=t.stateNode,ln(),F(pe),F(ie),Ii(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(gr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Oe!==null&&(bo(Oe),Oe=null))),Qo(e,t),le(t),null;case 5:Oi(t);var l=Ct(Xn.current);if(n=t.type,e!==null&&t.stateNode!=null)cc(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(y(166));return le(t),null}if(e=Ct(Be.current),gr(t)){r=t.stateNode,n=t.type;var o=t.memoizedProps;switch(r[Ue]=t,r[Kn]=o,e=(t.mode&1)!==0,n){case"dialog":D("cancel",r),D("close",r);break;case"iframe":case"object":case"embed":D("load",r);break;case"video":case"audio":for(l=0;l<Pn.length;l++)D(Pn[l],r);break;case"source":D("error",r);break;case"img":case"image":case"link":D("error",r),D("load",r);break;case"details":D("toggle",r);break;case"input":uu(r,o),D("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!o.multiple},D("invalid",r);break;case"textarea":au(r,o),D("invalid",r)}go(n,o),l=null;for(var i in o)if(o.hasOwnProperty(i)){var u=o[i];i==="children"?typeof u=="string"?r.textContent!==u&&(o.suppressHydrationWarning!==!0&&yr(r.textContent,u,e),l=["children",u]):typeof u=="number"&&r.textContent!==""+u&&(o.suppressHydrationWarning!==!0&&yr(r.textContent,u,e),l=["children",""+u]):Dn.hasOwnProperty(i)&&u!=null&&i==="onScroll"&&D("scroll",r)}switch(n){case"input":ar(r),su(r,o,!0);break;case"textarea":ar(r),cu(r);break;case"select":case"option":break;default:typeof o.onClick=="function"&&(r.onclick=Qr)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{i=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=Us(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=i.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=i.createElement(n,{is:r.is}):(e=i.createElement(n),n==="select"&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,n),e[Ue]=t,e[Kn]=r,ac(e,t,!1,!1),t.stateNode=e;e:{switch(i=wo(n,r),n){case"dialog":D("cancel",e),D("close",e),l=r;break;case"iframe":case"object":case"embed":D("load",e),l=r;break;case"video":case"audio":for(l=0;l<Pn.length;l++)D(Pn[l],e);l=r;break;case"source":D("error",e),l=r;break;case"img":case"image":case"link":D("error",e),D("load",e),l=r;break;case"details":D("toggle",e),l=r;break;case"input":uu(e,r),l=po(e,r),D("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=V({},r,{value:void 0}),D("invalid",e);break;case"textarea":au(e,r),l=vo(e,r),D("invalid",e);break;default:l=r}go(n,l),u=l;for(o in u)if(u.hasOwnProperty(o)){var s=u[o];o==="style"?Vs(e,s):o==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&$s(e,s)):o==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&Fn(e,s):typeof s=="number"&&Fn(e,""+s):o!=="suppressContentEditableWarning"&&o!=="suppressHydrationWarning"&&o!=="autoFocus"&&(Dn.hasOwnProperty(o)?s!=null&&o==="onScroll"&&D("scroll",e):s!=null&&ai(e,o,s,i))}switch(n){case"input":ar(e),su(e,r,!1);break;case"textarea":ar(e),cu(e);break;case"option":r.value!=null&&e.setAttribute("value",""+pt(r.value));break;case"select":e.multiple=!!r.multiple,o=r.value,o!=null?Xt(e,!!r.multiple,o,!1):r.defaultValue!=null&&Xt(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=Qr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return le(t),null;case 6:if(e&&t.stateNode!=null)fc(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(y(166));if(n=Ct(Xn.current),Ct(Be.current),gr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Ue]=t,(o=r.nodeValue!==n)&&(e=ge,e!==null))switch(e.tag){case 3:yr(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&yr(r.nodeValue,n,(e.mode&1)!==0)}o&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Ue]=t,t.stateNode=r}return le(t),null;case 13:if(F($),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(U&&ye!==null&&t.mode&1&&!(t.flags&128))La(),nn(),t.flags|=98560,o=!1;else if(o=gr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!o)throw Error(y(318));if(o=t.memoizedState,o=o!==null?o.dehydrated:null,!o)throw Error(y(317));o[Ue]=t}else nn(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;le(t),o=!1}else Oe!==null&&(bo(Oe),Oe=null),o=!0;if(!o)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||$.current&1?G===0&&(G=3):Qi())),t.updateQueue!==null&&(t.flags|=4),le(t),null);case 4:return ln(),Qo(e,t),e===null&&Hn(t.stateNode.containerInfo),le(t),null;case 10:return Li(t.type._context),le(t),null;case 17:return he(t.type)&&Kr(),le(t),null;case 19:if(F($),o=t.memoizedState,o===null)return le(t),null;if(r=(t.flags&128)!==0,i=o.rendering,i===null)if(r)kn(o,!1);else{if(G!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(i=br(e),i!==null){for(t.flags|=128,kn(o,!1),r=i.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)o=n,e=r,o.flags&=14680066,i=o.alternate,i===null?(o.childLanes=0,o.lanes=e,o.child=null,o.subtreeFlags=0,o.memoizedProps=null,o.memoizedState=null,o.updateQueue=null,o.dependencies=null,o.stateNode=null):(o.childLanes=i.childLanes,o.lanes=i.lanes,o.child=i.child,o.subtreeFlags=0,o.deletions=null,o.memoizedProps=i.memoizedProps,o.memoizedState=i.memoizedState,o.updateQueue=i.updateQueue,o.type=i.type,e=i.dependencies,o.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return M($,$.current&1|2),t.child}e=e.sibling}o.tail!==null&&Q()>un&&(t.flags|=128,r=!0,kn(o,!1),t.lanes=4194304)}else{if(!r)if(e=br(i),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),kn(o,!0),o.tail===null&&o.tailMode==="hidden"&&!i.alternate&&!U)return le(t),null}else 2*Q()-o.renderingStartTime>un&&n!==1073741824&&(t.flags|=128,r=!0,kn(o,!1),t.lanes=4194304);o.isBackwards?(i.sibling=t.child,t.child=i):(n=o.last,n!==null?n.sibling=i:t.child=i,o.last=i)}return o.tail!==null?(t=o.tail,o.rendering=t,o.tail=t.sibling,o.renderingStartTime=Q(),t.sibling=null,n=$.current,M($,r?n&1|2:n&1),t):(le(t),null);case 22:case 23:return Hi(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?ve&1073741824&&(le(t),t.subtreeFlags&6&&(t.flags|=8192)):le(t),null;case 24:return null;case 25:return null}throw Error(y(156,t.tag))}function Zd(e,t){switch(Ci(t),t.tag){case 1:return he(t.type)&&Kr(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return ln(),F(pe),F(ie),Ii(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return Oi(t),null;case 13:if(F($),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(y(340));nn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return F($),null;case 4:return ln(),null;case 10:return Li(t.type._context),null;case 22:case 23:return Hi(),null;case 24:return null;default:return null}}var kr=!1,oe=!1,Jd=typeof WeakSet=="function"?WeakSet:Set,k=null;function Kt(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){W(e,t,r)}else n.current=null}function Ko(e,t,n){try{n()}catch(r){W(e,t,r)}}var bu=!1;function qd(e,t){if(zo=Wr,e=ma(),Ei(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,o=r.focusNode;r=r.focusOffset;try{n.nodeType,o.nodeType}catch{n=null;break e}var i=0,u=-1,s=-1,a=0,h=0,m=e,p=null;t:for(;;){for(var w;m!==n||l!==0&&m.nodeType!==3||(u=i+l),m!==o||r!==0&&m.nodeType!==3||(s=i+r),m.nodeType===3&&(i+=m.nodeValue.length),(w=m.firstChild)!==null;)p=m,m=w;for(;;){if(m===e)break t;if(p===n&&++a===l&&(u=i),p===o&&++h===r&&(s=i),(w=m.nextSibling)!==null)break;m=p,p=m.parentNode}m=w}n=u===-1||s===-1?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Ro={focusedElem:e,selectionRange:n},Wr=!1,k=t;k!==null;)if(t=k,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,k=e;else for(;k!==null;){t=k;try{var S=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(S!==null){var g=S.memoizedProps,L=S.memoizedState,f=t.stateNode,c=f.getSnapshotBeforeUpdate(t.elementType===t.type?g:Re(t.type,g),L);f.__reactInternalSnapshotBeforeUpdate=c}break;case 3:var d=t.stateNode.containerInfo;d.nodeType===1?d.textContent="":d.nodeType===9&&d.documentElement&&d.removeChild(d.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(y(163))}}catch(v){W(t,t.return,v)}if(e=t.sibling,e!==null){e.return=t.return,k=e;break}k=t.return}return S=bu,bu=!1,S}function In(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var o=l.destroy;l.destroy=void 0,o!==void 0&&Ko(t,n,o)}l=l.next}while(l!==r)}}function yl(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Yo(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function dc(e){var t=e.alternate;t!==null&&(e.alternate=null,dc(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Ue],delete t[Kn],delete t[Io],delete t[jd],delete t[Md])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function pc(e){return e.tag===5||e.tag===3||e.tag===4}function es(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||pc(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Xo(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=Qr));else if(r!==4&&(e=e.child,e!==null))for(Xo(e,t,n),e=e.sibling;e!==null;)Xo(e,t,n),e=e.sibling}function Go(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(Go(e,t,n),e=e.sibling;e!==null;)Go(e,t,n),e=e.sibling}var ee=null,Te=!1;function Je(e,t,n){for(n=n.child;n!==null;)hc(e,t,n),n=n.sibling}function hc(e,t,n){if($e&&typeof $e.onCommitFiberUnmount=="function")try{$e.onCommitFiberUnmount(al,n)}catch{}switch(n.tag){case 5:oe||Kt(n,t);case 6:var r=ee,l=Te;ee=null,Je(e,t,n),ee=r,Te=l,ee!==null&&(Te?(e=ee,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):ee.removeChild(n.stateNode));break;case 18:ee!==null&&(Te?(e=ee,n=n.stateNode,e.nodeType===8?Ql(e.parentNode,n):e.nodeType===1&&Ql(e,n),Vn(e)):Ql(ee,n.stateNode));break;case 4:r=ee,l=Te,ee=n.stateNode.containerInfo,Te=!0,Je(e,t,n),ee=r,Te=l;break;case 0:case 11:case 14:case 15:if(!oe&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var o=l,i=o.destroy;o=o.tag,i!==void 0&&(o&2||o&4)&&Ko(n,t,i),l=l.next}while(l!==r)}Je(e,t,n);break;case 1:if(!oe&&(Kt(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(u){W(n,t,u)}Je(e,t,n);break;case 21:Je(e,t,n);break;case 22:n.mode&1?(oe=(r=oe)||n.memoizedState!==null,Je(e,t,n),oe=r):Je(e,t,n);break;default:Je(e,t,n)}}function ts(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Jd),t.forEach(function(r){var l=up.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function ze(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var o=e,i=t,u=i;e:for(;u!==null;){switch(u.tag){case 5:ee=u.stateNode,Te=!1;break e;case 3:ee=u.stateNode.containerInfo,Te=!0;break e;case 4:ee=u.stateNode.containerInfo,Te=!0;break e}u=u.return}if(ee===null)throw Error(y(160));hc(o,i,l),ee=null,Te=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(a){W(l,t,a)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)mc(t,e),t=t.sibling}function mc(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(ze(t,e),De(e),r&4){try{In(3,e,e.return),yl(3,e)}catch(g){W(e,e.return,g)}try{In(5,e,e.return)}catch(g){W(e,e.return,g)}}break;case 1:ze(t,e),De(e),r&512&&n!==null&&Kt(n,n.return);break;case 5:if(ze(t,e),De(e),r&512&&n!==null&&Kt(n,n.return),e.flags&32){var l=e.stateNode;try{Fn(l,"")}catch(g){W(e,e.return,g)}}if(r&4&&(l=e.stateNode,l!=null)){var o=e.memoizedProps,i=n!==null?n.memoizedProps:o,u=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{u==="input"&&o.type==="radio"&&o.name!=null&&Ds(l,o),wo(u,i);var a=wo(u,o);for(i=0;i<s.length;i+=2){var h=s[i],m=s[i+1];h==="style"?Vs(l,m):h==="dangerouslySetInnerHTML"?$s(l,m):h==="children"?Fn(l,m):ai(l,h,m,a)}switch(u){case"input":ho(l,o);break;case"textarea":Fs(l,o);break;case"select":var p=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!o.multiple;var w=o.value;w!=null?Xt(l,!!o.multiple,w,!1):p!==!!o.multiple&&(o.defaultValue!=null?Xt(l,!!o.multiple,o.defaultValue,!0):Xt(l,!!o.multiple,o.multiple?[]:"",!1))}l[Kn]=o}catch(g){W(e,e.return,g)}}break;case 6:if(ze(t,e),De(e),r&4){if(e.stateNode===null)throw Error(y(162));l=e.stateNode,o=e.memoizedProps;try{l.nodeValue=o}catch(g){W(e,e.return,g)}}break;case 3:if(ze(t,e),De(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{Vn(t.containerInfo)}catch(g){W(e,e.return,g)}break;case 4:ze(t,e),De(e);break;case 13:ze(t,e),De(e),l=e.child,l.flags&8192&&(o=l.memoizedState!==null,l.stateNode.isHidden=o,!o||l.alternate!==null&&l.alternate.memoizedState!==null||(Wi=Q())),r&4&&ts(e);break;case 22:if(h=n!==null&&n.memoizedState!==null,e.mode&1?(oe=(a=oe)||h,ze(t,e),oe=a):ze(t,e),De(e),r&8192){if(a=e.memoizedState!==null,(e.stateNode.isHidden=a)&&!h&&e.mode&1)for(k=e,h=e.child;h!==null;){for(m=k=h;k!==null;){switch(p=k,w=p.child,p.tag){case 0:case 11:case 14:case 15:In(4,p,p.return);break;case 1:Kt(p,p.return);var S=p.stateNode;if(typeof S.componentWillUnmount=="function"){r=p,n=p.return;try{t=r,S.props=t.memoizedProps,S.state=t.memoizedState,S.componentWillUnmount()}catch(g){W(r,n,g)}}break;case 5:Kt(p,p.return);break;case 22:if(p.memoizedState!==null){rs(m);continue}}w!==null?(w.return=p,k=w):rs(m)}h=h.sibling}e:for(h=null,m=e;;){if(m.tag===5){if(h===null){h=m;try{l=m.stateNode,a?(o=l.style,typeof o.setProperty=="function"?o.setProperty("display","none","important"):o.display="none"):(u=m.stateNode,s=m.memoizedProps.style,i=s!=null&&s.hasOwnProperty("display")?s.display:null,u.style.display=Bs("display",i))}catch(g){W(e,e.return,g)}}}else if(m.tag===6){if(h===null)try{m.stateNode.nodeValue=a?"":m.memoizedProps}catch(g){W(e,e.return,g)}}else if((m.tag!==22&&m.tag!==23||m.memoizedState===null||m===e)&&m.child!==null){m.child.return=m,m=m.child;continue}if(m===e)break e;for(;m.sibling===null;){if(m.return===null||m.return===e)break e;h===m&&(h=null),m=m.return}h===m&&(h=null),m.sibling.return=m.return,m=m.sibling}}break;case 19:ze(t,e),De(e),r&4&&ts(e);break;case 21:break;default:ze(t,e),De(e)}}function De(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(pc(n)){var r=n;break e}n=n.return}throw Error(y(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(Fn(l,""),r.flags&=-33);var o=es(e);Go(e,o,l);break;case 3:case 4:var i=r.stateNode.containerInfo,u=es(e);Xo(e,u,i);break;default:throw Error(y(161))}}catch(s){W(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function bd(e,t,n){k=e,vc(e)}function vc(e,t,n){for(var r=(e.mode&1)!==0;k!==null;){var l=k,o=l.child;if(l.tag===22&&r){var i=l.memoizedState!==null||kr;if(!i){var u=l.alternate,s=u!==null&&u.memoizedState!==null||oe;u=kr;var a=oe;if(kr=i,(oe=s)&&!a)for(k=l;k!==null;)i=k,s=i.child,i.tag===22&&i.memoizedState!==null?ls(l):s!==null?(s.return=i,k=s):ls(l);for(;o!==null;)k=o,vc(o),o=o.sibling;k=l,kr=u,oe=a}ns(e)}else l.subtreeFlags&8772&&o!==null?(o.return=l,k=o):ns(e)}}function ns(e){for(;k!==null;){var t=k;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:oe||yl(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!oe)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:Re(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var o=t.updateQueue;o!==null&&$u(t,o,r);break;case 3:var i=t.updateQueue;if(i!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}$u(t,i,n)}break;case 5:var u=t.stateNode;if(n===null&&t.flags&4){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var a=t.alternate;if(a!==null){var h=a.memoizedState;if(h!==null){var m=h.dehydrated;m!==null&&Vn(m)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(y(163))}oe||t.flags&512&&Yo(t)}catch(p){W(t,t.return,p)}}if(t===e){k=null;break}if(n=t.sibling,n!==null){n.return=t.return,k=n;break}k=t.return}}function rs(e){for(;k!==null;){var t=k;if(t===e){k=null;break}var n=t.sibling;if(n!==null){n.return=t.return,k=n;break}k=t.return}}function ls(e){for(;k!==null;){var t=k;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{yl(4,t)}catch(s){W(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){W(t,l,s)}}var o=t.return;try{Yo(t)}catch(s){W(t,o,s)}break;case 5:var i=t.return;try{Yo(t)}catch(s){W(t,i,s)}}}catch(s){W(t,t.return,s)}if(t===e){k=null;break}var u=t.sibling;if(u!==null){u.return=t.return,k=u;break}k=t.return}}var ep=Math.ceil,nl=Ze.ReactCurrentDispatcher,Bi=Ze.ReactCurrentOwner,_e=Ze.ReactCurrentBatchConfig,I=0,q=null,Y=null,te=0,ve=0,Yt=vt(0),G=0,qn=null,Tt=0,gl=0,Vi=0,jn=null,fe=null,Wi=0,un=1/0,Ve=null,rl=!1,Zo=null,ct=null,Er=!1,rt=null,ll=0,Mn=0,Jo=null,Ir=-1,jr=0;function se(){return I&6?Q():Ir!==-1?Ir:Ir=Q()}function ft(e){return e.mode&1?I&2&&te!==0?te&-te:Fd.transition!==null?(jr===0&&(jr=bs()),jr):(e=j,e!==0||(e=window.event,e=e===void 0?16:ia(e.type)),e):1}function je(e,t,n,r){if(50<Mn)throw Mn=0,Jo=null,Error(y(185));tr(e,n,r),(!(I&2)||e!==q)&&(e===q&&(!(I&2)&&(gl|=n),G===4&&tt(e,te)),me(e,r),n===1&&I===0&&!(t.mode&1)&&(un=Q()+500,hl&&yt()))}function me(e,t){var n=e.callbackNode;Ff(e,t);var r=Vr(e,e===q?te:0);if(r===0)n!==null&&pu(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&pu(n),t===1)e.tag===0?Dd(os.bind(null,e)):_a(os.bind(null,e)),Od(function(){!(I&6)&&yt()}),n=null;else{switch(ea(r)){case 1:n=hi;break;case 4:n=Js;break;case 16:n=Br;break;case 536870912:n=qs;break;default:n=Br}n=Cc(n,yc.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function yc(e,t){if(Ir=-1,jr=0,I&6)throw Error(y(327));var n=e.callbackNode;if(bt()&&e.callbackNode!==n)return null;var r=Vr(e,e===q?te:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=ol(e,r);else{t=r;var l=I;I|=2;var o=wc();(q!==e||te!==t)&&(Ve=null,un=Q()+500,_t(e,t));do try{rp();break}catch(u){gc(e,u)}while(1);Ni(),nl.current=o,I=l,Y!==null?t=0:(q=null,te=0,t=G)}if(t!==0){if(t===2&&(l=Co(e),l!==0&&(r=l,t=qo(e,l))),t===1)throw n=qn,_t(e,0),tt(e,r),me(e,Q()),n;if(t===6)tt(e,r);else{if(l=e.current.alternate,!(r&30)&&!tp(l)&&(t=ol(e,r),t===2&&(o=Co(e),o!==0&&(r=o,t=qo(e,o))),t===1))throw n=qn,_t(e,0),tt(e,r),me(e,Q()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(y(345));case 2:kt(e,fe,Ve);break;case 3:if(tt(e,r),(r&130023424)===r&&(t=Wi+500-Q(),10<t)){if(Vr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){se(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=Oo(kt.bind(null,e,fe,Ve),t);break}kt(e,fe,Ve);break;case 4:if(tt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var i=31-Ie(r);o=1<<i,i=t[i],i>l&&(l=i),r&=~o}if(r=l,r=Q()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*ep(r/1960))-r,10<r){e.timeoutHandle=Oo(kt.bind(null,e,fe,Ve),r);break}kt(e,fe,Ve);break;case 5:kt(e,fe,Ve);break;default:throw Error(y(329))}}}return me(e,Q()),e.callbackNode===n?yc.bind(null,e):null}function qo(e,t){var n=jn;return e.current.memoizedState.isDehydrated&&(_t(e,t).flags|=256),e=ol(e,t),e!==2&&(t=fe,fe=n,t!==null&&bo(t)),e}function bo(e){fe===null?fe=e:fe.push.apply(fe,e)}function tp(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],o=l.getSnapshot;l=l.value;try{if(!Me(o(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function tt(e,t){for(t&=~Vi,t&=~gl,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Ie(t),r=1<<n;e[n]=-1,t&=~r}}function os(e){if(I&6)throw Error(y(327));bt();var t=Vr(e,0);if(!(t&1))return me(e,Q()),null;var n=ol(e,t);if(e.tag!==0&&n===2){var r=Co(e);r!==0&&(t=r,n=qo(e,r))}if(n===1)throw n=qn,_t(e,0),tt(e,t),me(e,Q()),n;if(n===6)throw Error(y(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,kt(e,fe,Ve),me(e,Q()),null}function Ai(e,t){var n=I;I|=1;try{return e(t)}finally{I=n,I===0&&(un=Q()+500,hl&&yt())}}function Ot(e){rt!==null&&rt.tag===0&&!(I&6)&&bt();var t=I;I|=1;var n=_e.transition,r=j;try{if(_e.transition=null,j=1,e)return e()}finally{j=r,_e.transition=n,I=t,!(I&6)&&yt()}}function Hi(){ve=Yt.current,F(Yt)}function _t(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,Td(n)),Y!==null)for(n=Y.return;n!==null;){var r=n;switch(Ci(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&Kr();break;case 3:ln(),F(pe),F(ie),Ii();break;case 5:Oi(r);break;case 4:ln();break;case 13:F($);break;case 19:F($);break;case 10:Li(r.type._context);break;case 22:case 23:Hi()}n=n.return}if(q=e,Y=e=dt(e.current,null),te=ve=t,G=0,qn=null,Vi=gl=Tt=0,fe=jn=null,xt!==null){for(t=0;t<xt.length;t++)if(n=xt[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,o=n.pending;if(o!==null){var i=o.next;o.next=l,r.next=i}n.pending=r}xt=null}return e}function gc(e,t){do{var n=Y;try{if(Ni(),Rr.current=tl,el){for(var r=B.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}el=!1}if(Rt=0,J=X=B=null,On=!1,Gn=0,Bi.current=null,n===null||n.return===null){G=1,qn=t,Y=null;break}e:{var o=e,i=n.return,u=n,s=t;if(t=te,u.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var a=s,h=u,m=h.tag;if(!(h.mode&1)&&(m===0||m===11||m===15)){var p=h.alternate;p?(h.updateQueue=p.updateQueue,h.memoizedState=p.memoizedState,h.lanes=p.lanes):(h.updateQueue=null,h.memoizedState=null)}var w=Ku(i);if(w!==null){w.flags&=-257,Yu(w,i,u,o,t),w.mode&1&&Qu(o,a,t),t=w,s=a;var S=t.updateQueue;if(S===null){var g=new Set;g.add(s),t.updateQueue=g}else S.add(s);break e}else{if(!(t&1)){Qu(o,a,t),Qi();break e}s=Error(y(426))}}else if(U&&u.mode&1){var L=Ku(i);if(L!==null){!(L.flags&65536)&&(L.flags|=256),Yu(L,i,u,o,t),_i(on(s,u));break e}}o=s=on(s,u),G!==4&&(G=2),jn===null?jn=[o]:jn.push(o),o=i;do{switch(o.tag){case 3:o.flags|=65536,t&=-t,o.lanes|=t;var f=tc(o,s,t);Uu(o,f);break e;case 1:u=s;var c=o.type,d=o.stateNode;if(!(o.flags&128)&&(typeof c.getDerivedStateFromError=="function"||d!==null&&typeof d.componentDidCatch=="function"&&(ct===null||!ct.has(d)))){o.flags|=65536,t&=-t,o.lanes|=t;var v=nc(o,u,t);Uu(o,v);break e}}o=o.return}while(o!==null)}kc(n)}catch(E){t=E,Y===n&&n!==null&&(Y=n=n.return);continue}break}while(1)}function wc(){var e=nl.current;return nl.current=tl,e===null?tl:e}function Qi(){(G===0||G===3||G===2)&&(G=4),q===null||!(Tt&268435455)&&!(gl&268435455)||tt(q,te)}function ol(e,t){var n=I;I|=2;var r=wc();(q!==e||te!==t)&&(Ve=null,_t(e,t));do try{np();break}catch(l){gc(e,l)}while(1);if(Ni(),I=n,nl.current=r,Y!==null)throw Error(y(261));return q=null,te=0,G}function np(){for(;Y!==null;)Sc(Y)}function rp(){for(;Y!==null&&!Lf();)Sc(Y)}function Sc(e){var t=xc(e.alternate,e,ve);e.memoizedProps=e.pendingProps,t===null?kc(e):Y=t,Bi.current=null}function kc(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=Zd(n,t),n!==null){n.flags&=32767,Y=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{G=6,Y=null;return}}else if(n=Gd(n,t,ve),n!==null){Y=n;return}if(t=t.sibling,t!==null){Y=t;return}Y=t=e}while(t!==null);G===0&&(G=5)}function kt(e,t,n){var r=j,l=_e.transition;try{_e.transition=null,j=1,lp(e,t,n,r)}finally{_e.transition=l,j=r}return null}function lp(e,t,n,r){do bt();while(rt!==null);if(I&6)throw Error(y(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(y(177));e.callbackNode=null,e.callbackPriority=0;var o=n.lanes|n.childLanes;if(Uf(e,o),e===q&&(Y=q=null,te=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||Er||(Er=!0,Cc(Br,function(){return bt(),null})),o=(n.flags&15990)!==0,n.subtreeFlags&15990||o){o=_e.transition,_e.transition=null;var i=j;j=1;var u=I;I|=4,Bi.current=null,qd(e,n),mc(n,e),Cd(Ro),Wr=!!zo,Ro=zo=null,e.current=n,bd(n),zf(),I=u,j=i,_e.transition=o}else e.current=n;if(Er&&(Er=!1,rt=e,ll=l),o=e.pendingLanes,o===0&&(ct=null),Of(n.stateNode),me(e,Q()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(rl)throw rl=!1,e=Zo,Zo=null,e;return ll&1&&e.tag!==0&&bt(),o=e.pendingLanes,o&1?e===Jo?Mn++:(Mn=0,Jo=e):Mn=0,yt(),null}function bt(){if(rt!==null){var e=ea(ll),t=_e.transition,n=j;try{if(_e.transition=null,j=16>e?16:e,rt===null)var r=!1;else{if(e=rt,rt=null,ll=0,I&6)throw Error(y(331));var l=I;for(I|=4,k=e.current;k!==null;){var o=k,i=o.child;if(k.flags&16){var u=o.deletions;if(u!==null){for(var s=0;s<u.length;s++){var a=u[s];for(k=a;k!==null;){var h=k;switch(h.tag){case 0:case 11:case 15:In(8,h,o)}var m=h.child;if(m!==null)m.return=h,k=m;else for(;k!==null;){h=k;var p=h.sibling,w=h.return;if(dc(h),h===a){k=null;break}if(p!==null){p.return=w,k=p;break}k=w}}}var S=o.alternate;if(S!==null){var g=S.child;if(g!==null){S.child=null;do{var L=g.sibling;g.sibling=null,g=L}while(g!==null)}}k=o}}if(o.subtreeFlags&2064&&i!==null)i.return=o,k=i;else e:for(;k!==null;){if(o=k,o.flags&2048)switch(o.tag){case 0:case 11:case 15:In(9,o,o.return)}var f=o.sibling;if(f!==null){f.return=o.return,k=f;break e}k=o.return}}var c=e.current;for(k=c;k!==null;){i=k;var d=i.child;if(i.subtreeFlags&2064&&d!==null)d.return=i,k=d;else e:for(i=c;k!==null;){if(u=k,u.flags&2048)try{switch(u.tag){case 0:case 11:case 15:yl(9,u)}}catch(E){W(u,u.return,E)}if(u===i){k=null;break e}var v=u.sibling;if(v!==null){v.return=u.return,k=v;break e}k=u.return}}if(I=l,yt(),$e&&typeof $e.onPostCommitFiberRoot=="function")try{$e.onPostCommitFiberRoot(al,e)}catch{}r=!0}return r}finally{j=n,_e.transition=t}}return!1}function is(e,t,n){t=on(n,t),t=tc(e,t,1),e=at(e,t,1),t=se(),e!==null&&(tr(e,1,t),me(e,t))}function W(e,t,n){if(e.tag===3)is(e,e,n);else for(;t!==null;){if(t.tag===3){is(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(ct===null||!ct.has(r))){e=on(n,e),e=nc(t,e,1),t=at(t,e,1),e=se(),t!==null&&(tr(t,1,e),me(t,e));break}}t=t.return}}function op(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=se(),e.pingedLanes|=e.suspendedLanes&n,q===e&&(te&n)===n&&(G===4||G===3&&(te&130023424)===te&&500>Q()-Wi?_t(e,0):Vi|=n),me(e,t)}function Ec(e,t){t===0&&(e.mode&1?(t=dr,dr<<=1,!(dr&130023424)&&(dr=4194304)):t=1);var n=se();e=Xe(e,t),e!==null&&(tr(e,t,n),me(e,n))}function ip(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),Ec(e,n)}function up(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(y(314))}r!==null&&r.delete(t),Ec(e,n)}var xc;xc=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||pe.current)de=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return de=!1,Xd(e,t,n);de=!!(e.flags&131072)}else de=!1,U&&t.flags&1048576&&Pa(t,Gr,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;Or(e,t),e=t.pendingProps;var l=tn(t,ie.current);qt(t,n),l=Mi(null,t,r,e,l,n);var o=Di();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,he(r)?(o=!0,Yr(t)):o=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,Ri(t),l.updater=ml,t.stateNode=l,l._reactInternals=t,$o(t,r,e,n),t=Wo(null,t,r,!0,o,n)):(t.tag=0,U&&o&&xi(t),ue(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(Or(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=ap(r),e=Re(r,e),l){case 0:t=Vo(null,t,r,e,n);break e;case 1:t=Zu(null,t,r,e,n);break e;case 11:t=Xu(null,t,r,e,n);break e;case 14:t=Gu(null,t,r,Re(r.type,e),n);break e}throw Error(y(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Vo(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Zu(e,t,r,l,n);case 3:e:{if(ic(t),e===null)throw Error(y(387));r=t.pendingProps,o=t.memoizedState,l=o.element,Ra(e,t),qr(t,r,null,n);var i=t.memoizedState;if(r=i.element,o.isDehydrated)if(o={element:r,isDehydrated:!1,cache:i.cache,pendingSuspenseBoundaries:i.pendingSuspenseBoundaries,transitions:i.transitions},t.updateQueue.baseState=o,t.memoizedState=o,t.flags&256){l=on(Error(y(423)),t),t=Ju(e,t,r,n,l);break e}else if(r!==l){l=on(Error(y(424)),t),t=Ju(e,t,r,n,l);break e}else for(ye=st(t.stateNode.containerInfo.firstChild),ge=t,U=!0,Oe=null,n=ja(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(nn(),r===l){t=Ge(e,t,n);break e}ue(e,t,r,n)}t=t.child}return t;case 5:return Ma(t),e===null&&Do(t),r=t.type,l=t.pendingProps,o=e!==null?e.memoizedProps:null,i=l.children,To(r,l)?i=null:o!==null&&To(r,o)&&(t.flags|=32),oc(e,t),ue(e,t,i,n),t.child;case 6:return e===null&&Do(t),null;case 13:return uc(e,t,n);case 4:return Ti(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=rn(t,null,r,n):ue(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Xu(e,t,r,l,n);case 7:return ue(e,t,t.pendingProps,n),t.child;case 8:return ue(e,t,t.pendingProps.children,n),t.child;case 12:return ue(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,o=t.memoizedProps,i=l.value,M(Zr,r._currentValue),r._currentValue=i,o!==null)if(Me(o.value,i)){if(o.children===l.children&&!pe.current){t=Ge(e,t,n);break e}}else for(o=t.child,o!==null&&(o.return=t);o!==null;){var u=o.dependencies;if(u!==null){i=o.child;for(var s=u.firstContext;s!==null;){if(s.context===r){if(o.tag===1){s=Qe(-1,n&-n),s.tag=2;var a=o.updateQueue;if(a!==null){a=a.shared;var h=a.pending;h===null?s.next=s:(s.next=h.next,h.next=s),a.pending=s}}o.lanes|=n,s=o.alternate,s!==null&&(s.lanes|=n),Fo(o.return,n,t),u.lanes|=n;break}s=s.next}}else if(o.tag===10)i=o.type===t.type?null:o.child;else if(o.tag===18){if(i=o.return,i===null)throw Error(y(341));i.lanes|=n,u=i.alternate,u!==null&&(u.lanes|=n),Fo(i,n,t),i=o.sibling}else i=o.child;if(i!==null)i.return=o;else for(i=o;i!==null;){if(i===t){i=null;break}if(o=i.sibling,o!==null){o.return=i.return,i=o;break}i=i.return}o=i}ue(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,qt(t,n),l=Pe(l),r=r(l),t.flags|=1,ue(e,t,r,n),t.child;case 14:return r=t.type,l=Re(r,t.pendingProps),l=Re(r.type,l),Gu(e,t,r,l,n);case 15:return rc(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Or(e,t),t.tag=1,he(r)?(e=!0,Yr(t)):e=!1,qt(t,n),Oa(t,r,l),$o(t,r,l,n),Wo(null,t,r,!0,e,n);case 19:return sc(e,t,n);case 22:return lc(e,t,n)}throw Error(y(156,t.tag))};function Cc(e,t){return Zs(e,t)}function sp(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Ce(e,t,n,r){return new sp(e,t,n,r)}function Ki(e){return e=e.prototype,!(!e||!e.isReactComponent)}function ap(e){if(typeof e=="function")return Ki(e)?1:0;if(e!=null){if(e=e.$$typeof,e===fi)return 11;if(e===di)return 14}return 2}function dt(e,t){var n=e.alternate;return n===null?(n=Ce(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Mr(e,t,n,r,l,o){var i=2;if(r=e,typeof e=="function")Ki(e)&&(i=1);else if(typeof e=="string")i=5;else e:switch(e){case Ft:return Pt(n.children,l,o,t);case ci:i=8,l|=8;break;case so:return e=Ce(12,n,t,l|2),e.elementType=so,e.lanes=o,e;case ao:return e=Ce(13,n,t,l),e.elementType=ao,e.lanes=o,e;case co:return e=Ce(19,n,t,l),e.elementType=co,e.lanes=o,e;case Is:return wl(n,l,o,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case Ts:i=10;break e;case Os:i=9;break e;case fi:i=11;break e;case di:i=14;break e;case qe:i=16,r=null;break e}throw Error(y(130,e==null?e:typeof e,""))}return t=Ce(i,n,t,l),t.elementType=e,t.type=r,t.lanes=o,t}function Pt(e,t,n,r){return e=Ce(7,e,r,t),e.lanes=n,e}function wl(e,t,n,r){return e=Ce(22,e,r,t),e.elementType=Is,e.lanes=n,e.stateNode={isHidden:!1},e}function bl(e,t,n){return e=Ce(6,e,null,t),e.lanes=n,e}function eo(e,t,n){return t=Ce(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function cp(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=jl(0),this.expirationTimes=jl(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=jl(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Yi(e,t,n,r,l,o,i,u,s){return e=new cp(e,t,n,u,s),t===1?(t=1,o===!0&&(t|=8)):t=0,o=Ce(3,null,null,t),e.current=o,o.stateNode=e,o.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},Ri(o),e}function fp(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:Dt,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function _c(e){if(!e)return ht;e=e._reactInternals;e:{if(jt(e)!==e||e.tag!==1)throw Error(y(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(he(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(y(171))}if(e.tag===1){var n=e.type;if(he(n))return Ca(e,n,t)}return t}function Pc(e,t,n,r,l,o,i,u,s){return e=Yi(n,r,!0,e,l,o,i,u,s),e.context=_c(null),n=e.current,r=se(),l=ft(n),o=Qe(r,l),o.callback=t??null,at(n,o,l),e.current.lanes=l,tr(e,l,r),me(e,r),e}function Sl(e,t,n,r){var l=t.current,o=se(),i=ft(l);return n=_c(n),t.context===null?t.context=n:t.pendingContext=n,t=Qe(o,i),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=at(l,t,i),e!==null&&(je(e,l,i,o),zr(e,l,i)),i}function il(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function us(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Xi(e,t){us(e,t),(e=e.alternate)&&us(e,t)}function dp(){return null}var Nc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Gi(e){this._internalRoot=e}kl.prototype.render=Gi.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(y(409));Sl(e,t,null,null)};kl.prototype.unmount=Gi.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Ot(function(){Sl(null,e,null,null)}),t[Ye]=null}};function kl(e){this._internalRoot=e}kl.prototype.unstable_scheduleHydration=function(e){if(e){var t=ra();e={blockedOn:null,target:e,priority:t};for(var n=0;n<et.length&&t!==0&&t<et[n].priority;n++);et.splice(n,0,e),n===0&&oa(e)}};function Zi(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function El(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function ss(){}function pp(e,t,n,r,l){if(l){if(typeof r=="function"){var o=r;r=function(){var a=il(i);o.call(a)}}var i=Pc(t,r,e,0,null,!1,!1,"",ss);return e._reactRootContainer=i,e[Ye]=i.current,Hn(e.nodeType===8?e.parentNode:e),Ot(),i}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var u=r;r=function(){var a=il(s);u.call(a)}}var s=Yi(e,0,!1,null,null,!1,!1,"",ss);return e._reactRootContainer=s,e[Ye]=s.current,Hn(e.nodeType===8?e.parentNode:e),Ot(function(){Sl(t,s,n,r)}),s}function xl(e,t,n,r,l){var o=n._reactRootContainer;if(o){var i=o;if(typeof l=="function"){var u=l;l=function(){var s=il(i);u.call(s)}}Sl(t,i,e,l)}else i=pp(n,t,e,l,r);return il(i)}ta=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=_n(t.pendingLanes);n!==0&&(mi(t,n|1),me(t,Q()),!(I&6)&&(un=Q()+500,yt()))}break;case 13:Ot(function(){var r=Xe(e,1);if(r!==null){var l=se();je(r,e,1,l)}}),Xi(e,1)}};vi=function(e){if(e.tag===13){var t=Xe(e,134217728);if(t!==null){var n=se();je(t,e,134217728,n)}Xi(e,134217728)}};na=function(e){if(e.tag===13){var t=ft(e),n=Xe(e,t);if(n!==null){var r=se();je(n,e,t,r)}Xi(e,t)}};ra=function(){return j};la=function(e,t){var n=j;try{return j=e,t()}finally{j=n}};ko=function(e,t,n){switch(t){case"input":if(ho(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=pl(r);if(!l)throw Error(y(90));Ms(r),ho(r,l)}}}break;case"textarea":Fs(e,n);break;case"select":t=n.value,t!=null&&Xt(e,!!n.multiple,t,!1)}};Hs=Ai;Qs=Ot;var hp={usingClientEntryPoint:!1,Events:[rr,Vt,pl,Ws,As,Ai]},En={findFiberByHostInstance:Et,bundleType:0,version:"18.2.0",rendererPackageName:"react-dom"},mp={bundleType:En.bundleType,version:En.version,rendererPackageName:En.rendererPackageName,rendererConfig:En.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:Ze.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=Xs(e),e===null?null:e.stateNode},findFiberByHostInstance:En.findFiberByHostInstance||dp,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.2.0-next-9e3b772b8-20220608"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var xr=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!xr.isDisabled&&xr.supportsFiber)try{al=xr.inject(mp),$e=xr}catch{}}Se.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=hp;Se.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Zi(t))throw Error(y(200));return fp(e,t,null,n)};Se.createRoot=function(e,t){if(!Zi(e))throw Error(y(299));var n=!1,r="",l=Nc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Yi(e,1,!1,null,null,n,!1,r,l),e[Ye]=t.current,Hn(e.nodeType===8?e.parentNode:e),new Gi(t)};Se.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(y(188)):(e=Object.keys(e).join(","),Error(y(268,e)));return e=Xs(t),e=e===null?null:e.stateNode,e};Se.flushSync=function(e){return Ot(e)};Se.hydrate=function(e,t,n){if(!El(t))throw Error(y(200));return xl(null,e,t,!0,n)};Se.hydrateRoot=function(e,t,n){if(!Zi(e))throw Error(y(405));var r=n!=null&&n.hydratedSources||null,l=!1,o="",i=Nc;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(o=n.identifierPrefix),n.onRecoverableError!==void 0&&(i=n.onRecoverableError)),t=Pc(t,null,e,1,n??null,l,!1,o,i),e[Ye]=t.current,Hn(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new kl(t)};Se.render=function(e,t,n){if(!El(t))throw Error(y(200));return xl(null,e,t,!1,n)};Se.unmountComponentAtNode=function(e){if(!El(e))throw Error(y(40));return e._reactRootContainer?(Ot(function(){xl(null,null,e,!1,function(){e._reactRootContainer=null,e[Ye]=null})}),!0):!1};Se.unstable_batchedUpdates=Ai;Se.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!El(n))throw Error(y(200));if(e==null||e._reactInternals===void 0)throw Error(y(38));return xl(e,t,n,!1,r)};Se.version="18.2.0-next-9e3b772b8-20220608";(function(e){function t(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(t)}catch(n){console.error(n)}}t(),e.exports=Se})(df);var as=oo;lo.createRoot=as.createRoot,lo.hydrateRoot=as.hydrateRoot;const vp="modulepreload",yp=function(e){return"/service-worker-practice/"+e},cs={},Ji=function(t,n,r){if(!n||n.length===0)return t();const l=document.getElementsByTagName("link");return Promise.all(n.map(o=>{if(o=yp(o),o in cs)return;cs[o]=!0;const i=o.endsWith(".css"),u=i?'[rel="stylesheet"]':"";if(!!r)for(let h=l.length-1;h>=0;h--){const m=l[h];if(m.href===o&&(!i||m.rel==="stylesheet"))return}else if(document.querySelector(`link[href="${o}"]${u}`))return;const a=document.createElement("link");if(a.rel=i?"stylesheet":vp,i||(a.as="script",a.crossOrigin=""),a.href=o,document.head.appendChild(a),i)return new Promise((h,m)=>{a.addEventListener("load",h),a.addEventListener("error",()=>m(new Error(`Unable to preload CSS for ${o}`)))})})).then(()=>t())};/**
+ * @remix-run/router v1.5.0
+ *
+ * Copyright (c) Remix Software Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.md file in the root directory of this source tree.
+ *
+ * @license MIT
+ */function bn(){return bn=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},bn.apply(this,arguments)}var lt;(function(e){e.Pop="POP",e.Push="PUSH",e.Replace="REPLACE"})(lt||(lt={}));const fs="popstate";function gp(e){e===void 0&&(e={});function t(r,l){let{pathname:o,search:i,hash:u}=r.location;return ei("",{pathname:o,search:i,hash:u},l.state&&l.state.usr||null,l.state&&l.state.key||"default")}function n(r,l){return typeof l=="string"?l:Lc(l)}return Sp(t,n,null,e)}function b(e,t){if(e===!1||e===null||typeof e>"u")throw new Error(t)}function qi(e,t){if(!e){typeof console<"u"&&console.warn(t);try{throw new Error(t)}catch{}}}function wp(){return Math.random().toString(36).substr(2,8)}function ds(e,t){return{usr:e.state,key:e.key,idx:t}}function ei(e,t,n,r){return n===void 0&&(n=null),bn({pathname:typeof e=="string"?e:e.pathname,search:"",hash:""},typeof t=="string"?fn(t):t,{state:n,key:t&&t.key||r||wp()})}function Lc(e){let{pathname:t="/",search:n="",hash:r=""}=e;return n&&n!=="?"&&(t+=n.charAt(0)==="?"?n:"?"+n),r&&r!=="#"&&(t+=r.charAt(0)==="#"?r:"#"+r),t}function fn(e){let t={};if(e){let n=e.indexOf("#");n>=0&&(t.hash=e.substr(n),e=e.substr(0,n));let r=e.indexOf("?");r>=0&&(t.search=e.substr(r),e=e.substr(0,r)),e&&(t.pathname=e)}return t}function Sp(e,t,n,r){r===void 0&&(r={});let{window:l=document.defaultView,v5Compat:o=!1}=r,i=l.history,u=lt.Pop,s=null,a=h();a==null&&(a=0,i.replaceState(bn({},i.state,{idx:a}),""));function h(){return(i.state||{idx:null}).idx}function m(){u=lt.Pop;let L=h(),f=L==null?null:L-a;a=L,s&&s({action:u,location:g.location,delta:f})}function p(L,f){u=lt.Push;let c=ei(g.location,L,f);n&&n(c,L),a=h()+1;let d=ds(c,a),v=g.createHref(c);try{i.pushState(d,"",v)}catch{l.location.assign(v)}o&&s&&s({action:u,location:g.location,delta:1})}function w(L,f){u=lt.Replace;let c=ei(g.location,L,f);n&&n(c,L),a=h();let d=ds(c,a),v=g.createHref(c);i.replaceState(d,"",v),o&&s&&s({action:u,location:g.location,delta:0})}function S(L){let f=l.location.origin!=="null"?l.location.origin:l.location.href,c=typeof L=="string"?L:Lc(L);return b(f,"No window.location.(origin|href) available to create URL for href: "+c),new URL(c,f)}let g={get action(){return u},get location(){return e(l,i)},listen(L){if(s)throw new Error("A history only accepts one active listener");return l.addEventListener(fs,m),s=L,()=>{l.removeEventListener(fs,m),s=null}},createHref(L){return t(l,L)},createURL:S,encodeLocation(L){let f=S(L);return{pathname:f.pathname,search:f.search,hash:f.hash}},push:p,replace:w,go(L){return i.go(L)}};return g}var ps;(function(e){e.data="data",e.deferred="deferred",e.redirect="redirect",e.error="error"})(ps||(ps={}));function kp(e,t,n){n===void 0&&(n="/");let r=typeof t=="string"?fn(t):t,l=Tc(r.pathname||"/",n);if(l==null)return null;let o=zc(e);Ep(o);let i=null;for(let u=0;i==null&&u<o.length;++u)i=Tp(o[u],jp(l));return i}function zc(e,t,n,r){t===void 0&&(t=[]),n===void 0&&(n=[]),r===void 0&&(r="");let l=(o,i,u)=>{let s={relativePath:u===void 0?o.path||"":u,caseSensitive:o.caseSensitive===!0,childrenIndex:i,route:o};s.relativePath.startsWith("/")&&(b(s.relativePath.startsWith(r),'Absolute route path "'+s.relativePath+'" nested under path '+('"'+r+'" is not valid. An absolute child route path ')+"must start with the combined path of all its parent routes."),s.relativePath=s.relativePath.slice(r.length));let a=Nt([r,s.relativePath]),h=n.concat(s);o.children&&o.children.length>0&&(b(o.index!==!0,"Index routes must not have child routes. Please remove "+('all child routes from route path "'+a+'".')),zc(o.children,t,h,a)),!(o.path==null&&!o.index)&&t.push({path:a,score:zp(a,o.index),routesMeta:h})};return e.forEach((o,i)=>{var u;if(o.path===""||!((u=o.path)!=null&&u.includes("?")))l(o,i);else for(let s of Rc(o.path))l(o,i,s)}),t}function Rc(e){let t=e.split("/");if(t.length===0)return[];let[n,...r]=t,l=n.endsWith("?"),o=n.replace(/\?$/,"");if(r.length===0)return l?[o,""]:[o];let i=Rc(r.join("/")),u=[];return u.push(...i.map(s=>s===""?o:[o,s].join("/"))),l&&u.push(...i),u.map(s=>e.startsWith("/")&&s===""?"/":s)}function Ep(e){e.sort((t,n)=>t.score!==n.score?n.score-t.score:Rp(t.routesMeta.map(r=>r.childrenIndex),n.routesMeta.map(r=>r.childrenIndex)))}const xp=/^:\w+$/,Cp=3,_p=2,Pp=1,Np=10,Lp=-2,hs=e=>e==="*";function zp(e,t){let n=e.split("/"),r=n.length;return n.some(hs)&&(r+=Lp),t&&(r+=_p),n.filter(l=>!hs(l)).reduce((l,o)=>l+(xp.test(o)?Cp:o===""?Pp:Np),r)}function Rp(e,t){return e.length===t.length&&e.slice(0,-1).every((r,l)=>r===t[l])?e[e.length-1]-t[t.length-1]:0}function Tp(e,t){let{routesMeta:n}=e,r={},l="/",o=[];for(let i=0;i<n.length;++i){let u=n[i],s=i===n.length-1,a=l==="/"?t:t.slice(l.length)||"/",h=Op({path:u.relativePath,caseSensitive:u.caseSensitive,end:s},a);if(!h)return null;Object.assign(r,h.params);let m=u.route;o.push({params:r,pathname:Nt([l,h.pathname]),pathnameBase:Bp(Nt([l,h.pathnameBase])),route:m}),h.pathnameBase!=="/"&&(l=Nt([l,h.pathnameBase]))}return o}function Op(e,t){typeof e=="string"&&(e={path:e,caseSensitive:!1,end:!0});let[n,r]=Ip(e.path,e.caseSensitive,e.end),l=t.match(n);if(!l)return null;let o=l[0],i=o.replace(/(.)\/+$/,"$1"),u=l.slice(1);return{params:r.reduce((a,h,m)=>{if(h==="*"){let p=u[m]||"";i=o.slice(0,o.length-p.length).replace(/(.)\/+$/,"$1")}return a[h]=Mp(u[m]||"",h),a},{}),pathname:o,pathnameBase:i,pattern:e}}function Ip(e,t,n){t===void 0&&(t=!1),n===void 0&&(n=!0),qi(e==="*"||!e.endsWith("*")||e.endsWith("/*"),'Route path "'+e+'" will be treated as if it were '+('"'+e.replace(/\*$/,"/*")+'" because the `*` character must ')+"always follow a `/` in the pattern. To get rid of this warning, "+('please change the route path to "'+e.replace(/\*$/,"/*")+'".'));let r=[],l="^"+e.replace(/\/*\*?$/,"").replace(/^\/*/,"/").replace(/[\\.*+^$?{}|()[\]]/g,"\\$&").replace(/\/:(\w+)/g,(i,u)=>(r.push(u),"/([^\\/]+)"));return e.endsWith("*")?(r.push("*"),l+=e==="*"||e==="/*"?"(.*)$":"(?:\\/(.+)|\\/*)$"):n?l+="\\/*$":e!==""&&e!=="/"&&(l+="(?:(?=\\/|$))"),[new RegExp(l,t?void 0:"i"),r]}function jp(e){try{return decodeURI(e)}catch(t){return qi(!1,'The URL path "'+e+'" could not be decoded because it is is a malformed URL segment. This is probably due to a bad percent '+("encoding ("+t+").")),e}}function Mp(e,t){try{return decodeURIComponent(e)}catch(n){return qi(!1,'The value for the URL param "'+t+'" will not be decoded because'+(' the string "'+e+'" is a malformed URL segment. This is probably')+(" due to a bad percent encoding ("+n+").")),e}}function Tc(e,t){if(t==="/")return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let n=t.endsWith("/")?t.length-1:t.length,r=e.charAt(n);return r&&r!=="/"?null:e.slice(n)||"/"}function Dp(e,t){t===void 0&&(t="/");let{pathname:n,search:r="",hash:l=""}=typeof e=="string"?fn(e):e;return{pathname:n?n.startsWith("/")?n:Fp(n,t):t,search:Vp(r),hash:Wp(l)}}function Fp(e,t){let n=t.replace(/\/+$/,"").split("/");return e.split("/").forEach(l=>{l===".."?n.length>1&&n.pop():l!=="."&&n.push(l)}),n.length>1?n.join("/"):"/"}function to(e,t,n,r){return"Cannot include a '"+e+"' character in a manually specified "+("`to."+t+"` field ["+JSON.stringify(r)+"].  Please separate it out to the ")+("`to."+n+"` field. Alternatively you may provide the full path as ")+'a string in <Link to="..."> and the router will parse it for you.'}function Up(e){return e.filter((t,n)=>n===0||t.route.path&&t.route.path.length>0)}function $p(e,t,n,r){r===void 0&&(r=!1);let l;typeof e=="string"?l=fn(e):(l=bn({},e),b(!l.pathname||!l.pathname.includes("?"),to("?","pathname","search",l)),b(!l.pathname||!l.pathname.includes("#"),to("#","pathname","hash",l)),b(!l.search||!l.search.includes("#"),to("#","search","hash",l)));let o=e===""||l.pathname==="",i=o?"/":l.pathname,u;if(r||i==null)u=n;else{let m=t.length-1;if(i.startsWith("..")){let p=i.split("/");for(;p[0]==="..";)p.shift(),m-=1;l.pathname=p.join("/")}u=m>=0?t[m]:"/"}let s=Dp(l,u),a=i&&i!=="/"&&i.endsWith("/"),h=(o||i===".")&&n.endsWith("/");return!s.pathname.endsWith("/")&&(a||h)&&(s.pathname+="/"),s}const Nt=e=>e.join("/").replace(/\/\/+/g,"/"),Bp=e=>e.replace(/\/+$/,"").replace(/^\/*/,"/"),Vp=e=>!e||e==="?"?"":e.startsWith("?")?e:"?"+e,Wp=e=>!e||e==="#"?"":e.startsWith("#")?e:"#"+e;function Ap(e){return e!=null&&typeof e.status=="number"&&typeof e.statusText=="string"&&typeof e.internal=="boolean"&&"data"in e}/**
+ * React Router v6.10.0
+ *
+ * Copyright (c) Remix Software Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.md file in the root directory of this source tree.
+ *
+ * @license MIT
+ */function Hp(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}const Qp=typeof Object.is=="function"?Object.is:Hp,{useState:Kp,useEffect:Yp,useLayoutEffect:Xp,useDebugValue:Gp}=ro;function Zp(e,t,n){const r=t(),[{inst:l},o]=Kp({inst:{value:r,getSnapshot:t}});return Xp(()=>{l.value=r,l.getSnapshot=t,no(l)&&o({inst:l})},[e,r,t]),Yp(()=>(no(l)&&o({inst:l}),e(()=>{no(l)&&o({inst:l})})),[e]),Gp(r),r}function no(e){const t=e.getSnapshot,n=e.value;try{const r=t();return!Qp(n,r)}catch{return!0}}function Jp(e,t,n){return t()}const qp=typeof window<"u"&&typeof window.document<"u"&&typeof window.document.createElement<"u",bp=!qp,eh=bp?Jp:Zp;"useSyncExternalStore"in ro&&(e=>e.useSyncExternalStore)(ro);const Oc=C.createContext(null),Ic=C.createContext(null),bi=C.createContext(null),Cl=C.createContext(null),dn=C.createContext({outlet:null,matches:[]}),jc=C.createContext(null);function ti(){return ti=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},ti.apply(this,arguments)}function _l(){return C.useContext(Cl)!=null}function Mc(){return _l()||b(!1),C.useContext(Cl).location}function th(){_l()||b(!1);let{basename:e,navigator:t}=C.useContext(bi),{matches:n}=C.useContext(dn),{pathname:r}=Mc(),l=JSON.stringify(Up(n).map(u=>u.pathnameBase)),o=C.useRef(!1);return C.useEffect(()=>{o.current=!0}),C.useCallback(function(u,s){if(s===void 0&&(s={}),!o.current)return;if(typeof u=="number"){t.go(u);return}let a=$p(u,JSON.parse(l),r,s.relative==="path");e!=="/"&&(a.pathname=a.pathname==="/"?e:Nt([e,a.pathname])),(s.replace?t.replace:t.push)(a,s.state,s)},[e,t,l,r])}const nh=C.createContext(null);function rh(e){let t=C.useContext(dn).outlet;return t&&C.createElement(nh.Provider,{value:e},t)}function lh(e,t){_l()||b(!1);let{navigator:n}=C.useContext(bi),r=C.useContext(Ic),{matches:l}=C.useContext(dn),o=l[l.length-1],i=o?o.params:{};o&&o.pathname;let u=o?o.pathnameBase:"/";o&&o.route;let s=Mc(),a;if(t){var h;let g=typeof t=="string"?fn(t):t;u==="/"||(h=g.pathname)!=null&&h.startsWith(u)||b(!1),a=g}else a=s;let m=a.pathname||"/",p=u==="/"?m:m.slice(u.length)||"/",w=kp(e,{pathname:p}),S=sh(w&&w.map(g=>Object.assign({},g,{params:Object.assign({},i,g.params),pathname:Nt([u,n.encodeLocation?n.encodeLocation(g.pathname).pathname:g.pathname]),pathnameBase:g.pathnameBase==="/"?u:Nt([u,n.encodeLocation?n.encodeLocation(g.pathnameBase).pathname:g.pathnameBase])})),l,r||void 0);return t&&S?C.createElement(Cl.Provider,{value:{location:ti({pathname:"/",search:"",hash:"",state:null,key:"default"},a),navigationType:lt.Pop}},S):S}function oh(){let e=dh(),t=Ap(e)?e.status+" "+e.statusText:e instanceof Error?e.message:JSON.stringify(e),n=e instanceof Error?e.stack:null,l={padding:"0.5rem",backgroundColor:"rgba(200,200,200, 0.5)"},o=null;return C.createElement(C.Fragment,null,C.createElement("h2",null,"Unexpected Application Error!"),C.createElement("h3",{style:{fontStyle:"italic"}},t),n?C.createElement("pre",{style:l},n):null,o)}class ih extends C.Component{constructor(t){super(t),this.state={location:t.location,error:t.error}}static getDerivedStateFromError(t){return{error:t}}static getDerivedStateFromProps(t,n){return n.location!==t.location?{error:t.error,location:t.location}:{error:t.error||n.error,location:n.location}}componentDidCatch(t,n){console.error("React Router caught the following error during render",t,n)}render(){return this.state.error?C.createElement(dn.Provider,{value:this.props.routeContext},C.createElement(jc.Provider,{value:this.state.error,children:this.props.component})):this.props.children}}function uh(e){let{routeContext:t,match:n,children:r}=e,l=C.useContext(Oc);return l&&l.static&&l.staticContext&&(n.route.errorElement||n.route.ErrorBoundary)&&(l.staticContext._deepestRenderedBoundaryId=n.route.id),C.createElement(dn.Provider,{value:t},r)}function sh(e,t,n){if(t===void 0&&(t=[]),e==null)if(n!=null&&n.errors)e=n.matches;else return null;let r=e,l=n==null?void 0:n.errors;if(l!=null){let o=r.findIndex(i=>i.route.id&&(l==null?void 0:l[i.route.id]));o>=0||b(!1),r=r.slice(0,Math.min(r.length,o+1))}return r.reduceRight((o,i,u)=>{let s=i.route.id?l==null?void 0:l[i.route.id]:null,a=null;n&&(i.route.ErrorBoundary?a=C.createElement(i.route.ErrorBoundary,null):i.route.errorElement?a=i.route.errorElement:a=C.createElement(oh,null));let h=t.concat(r.slice(0,u+1)),m=()=>{let p=o;return s?p=a:i.route.Component?p=C.createElement(i.route.Component,null):i.route.element&&(p=i.route.element),C.createElement(uh,{match:i,routeContext:{outlet:o,matches:h},children:p})};return n&&(i.route.ErrorBoundary||i.route.errorElement||u===0)?C.createElement(ih,{location:n.location,component:a,error:s,children:m(),routeContext:{outlet:null,matches:h}}):m()},null)}var ms;(function(e){e.UseBlocker="useBlocker",e.UseRevalidator="useRevalidator"})(ms||(ms={}));var ul;(function(e){e.UseBlocker="useBlocker",e.UseLoaderData="useLoaderData",e.UseActionData="useActionData",e.UseRouteError="useRouteError",e.UseNavigation="useNavigation",e.UseRouteLoaderData="useRouteLoaderData",e.UseMatches="useMatches",e.UseRevalidator="useRevalidator"})(ul||(ul={}));function ah(e){let t=C.useContext(Ic);return t||b(!1),t}function ch(e){let t=C.useContext(dn);return t||b(!1),t}function fh(e){let t=ch(),n=t.matches[t.matches.length-1];return n.route.id||b(!1),n.route.id}function dh(){var e;let t=C.useContext(jc),n=ah(ul.UseRouteError),r=fh(ul.UseRouteError);return t||((e=n.errors)==null?void 0:e[r])}function ph(e){return rh(e.context)}function Dr(e){b(!1)}function hh(e){let{basename:t="/",children:n=null,location:r,navigationType:l=lt.Pop,navigator:o,static:i=!1}=e;_l()&&b(!1);let u=t.replace(/^\/*/,"/"),s=C.useMemo(()=>({basename:u,navigator:o,static:i}),[u,o,i]);typeof r=="string"&&(r=fn(r));let{pathname:a="/",search:h="",hash:m="",state:p=null,key:w="default"}=r,S=C.useMemo(()=>{let g=Tc(a,u);return g==null?null:{location:{pathname:g,search:h,hash:m,state:p,key:w},navigationType:l}},[u,a,h,m,p,w,l]);return S==null?null:C.createElement(bi.Provider,{value:s},C.createElement(Cl.Provider,{children:n,value:S}))}function mh(e){let{children:t,location:n}=e,r=C.useContext(Oc),l=r&&!t?r.router.routes:ni(t);return lh(l,n)}var vs;(function(e){e[e.pending=0]="pending",e[e.success=1]="success",e[e.error=2]="error"})(vs||(vs={}));new Promise(()=>{});function ni(e,t){t===void 0&&(t=[]);let n=[];return C.Children.forEach(e,(r,l)=>{if(!C.isValidElement(r))return;let o=[...t,l];if(r.type===C.Fragment){n.push.apply(n,ni(r.props.children,o));return}r.type!==Dr&&b(!1),!r.props.index||!r.props.children||b(!1);let i={id:r.props.id||o.join("-"),caseSensitive:r.props.caseSensitive,element:r.props.element,Component:r.props.Component,index:r.props.index,path:r.props.path,loader:r.props.loader,action:r.props.action,errorElement:r.props.errorElement,ErrorBoundary:r.props.ErrorBoundary,hasErrorBoundary:r.props.ErrorBoundary!=null||r.props.errorElement!=null,shouldRevalidate:r.props.shouldRevalidate,handle:r.props.handle,lazy:r.props.lazy};r.props.children&&(i.children=ni(r.props.children,o)),n.push(i)}),n}/**
+ * React Router DOM v6.10.0
+ *
+ * Copyright (c) Remix Software Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE.md file in the root directory of this source tree.
+ *
+ * @license MIT
+ */function vh(e){let{basename:t,children:n,window:r}=e,l=C.useRef();l.current==null&&(l.current=gp({window:r,v5Compat:!0}));let o=l.current,[i,u]=C.useState({action:o.action,location:o.location});return C.useLayoutEffect(()=>o.listen(u),[o]),C.createElement(hh,{basename:t,children:n,location:i.location,navigationType:i.action,navigator:o})}var ys;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmitImpl="useSubmitImpl",e.UseFetcher="useFetcher"})(ys||(ys={}));var gs;(function(e){e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(gs||(gs={}));const yh=()=>{const e=th();return t=>new Promise(n=>{C.startTransition(()=>{e(t),n()})})},Dc=()=>{const e=yh(),t=[{text:"top",onClick:()=>{e("/")}},{text:"playground",onClick:()=>{e("/playground")}}];return K.jsxs("div",{children:[K.jsx("nav",{children:t.map(n=>K.jsx("button",{onClick:n.onClick,children:n.text},n.text))}),K.jsx("div",{children:K.jsx(C.Suspense,{fallback:K.jsx("p",{children:"Loading..."}),children:K.jsx(ph,{})})})]})};Dc.displayName="BasicLayout";const gh=C.lazy(()=>Ji(()=>import("./Top-d2d68f9c.js"),[])),wh=C.lazy(()=>Ji(()=>import("./Playground-2de2d3cd.js"),[])),Fc=()=>K.jsx(vh,{basename:"/service-worker-practice",children:K.jsx(mh,{children:K.jsxs(Dr,{path:"/",element:K.jsx(Dc,{}),children:[K.jsx(Dr,{path:"/",element:K.jsx(gh,{})}),K.jsx(Dr,{path:"/playground",element:K.jsx(wh,{})})]})})});Fc.displayName="AppRouter";const Uc=()=>K.jsx(Fc,{});Uc.displayName="App";function Sh(e={}){const{immediate:t=!1,onNeedRefresh:n,onOfflineReady:r,onRegistered:l,onRegisteredSW:o,onRegisterError:i}=e;let u,s;const a=async(m=!0)=>{await s};async function h(){if("serviceWorker"in navigator){const{Workbox:m}=await Ji(()=>import("./workbox-window.prod.es5-dc90f814.js"),[]);u=new m("/service-worker-practice/sw.js",{scope:"/service-worker-practice/",type:"classic"}),u.addEventListener("activated",p=>{(p.isUpdate||p.isExternal)&&window.location.reload()}),u.addEventListener("installed",p=>{p.isUpdate||r==null||r()}),u.register({immediate:t}).then(p=>{o?o("/service-worker-practice/sw.js",p):l==null||l(p)}).catch(p=>{i==null||i(p)})}}return s=h(),a}lo.createRoot(document.getElementById("root")).render(K.jsx(Ps.StrictMode,{children:K.jsx(Uc,{})}));Sh();const kh=["TypeError: Failed to fetch dynamically imported module","Failed to load module script: Expected a JavaScript module script"].map(e=>new RegExp(e));window.addEventListener("error",e=>{kh.map(n=>n.test(e.message)).some(n=>n)&&alert("Failed to get resource.")});export{Ps as W,K as j,C as r};
