+++ docs/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/service-worker-practice/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + React + TS</title>
-    <script type="module" crossorigin src="/service-worker-practice/assets/index-10388f2a.js"></script>
+    <script type="module" crossorigin src="/service-worker-practice/assets/index-8bd2be33.js"></script>
   <link rel="manifest" href="/service-worker-practice/manifest.webmanifest"></head>
   <body>
     <div id="root"></div>
