+++ src/pages/Playground.tsx
@@ -8,7 +8,7 @@ const Playground: React.FC = () => {
   };
   return (
     <div>
-      <h1>Playground</h1>
+      <h1>Playground V2</h1>
       <div style={{ height:"50vh", width: "100%" }}>
         <MonacoEditor {...monacoEditorProps} />
       </div>
